% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{add_juvenile_faceting}
\alias{add_juvenile_faceting}
\title{Helper function to add faceting to juvenile plots}
\usage{
add_juvenile_faceting(p, facet_var, use_facet_wrap, facet_cols, fill_var)
}
\description{
Helper function to add faceting to juvenile plots
}
\keyword{internal}
