% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{add_juvenile_geoms}
\alias{add_juvenile_geoms}
\title{Helper function to add geoms to juvenile plots}
\usage{
add_juvenile_geoms(
  p,
  plot_type,
  bar_position,
  fill_var,
  use_facet_wrap,
  group_by_type
)
}
\description{
Helper function to add geoms to juvenile plots
}
\keyword{internal}
