% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{add_juvenile_reference_line}
\alias{add_juvenile_reference_line}
\title{Helper function to add reference line to juvenile plots}
\usage{
add_juvenile_reference_line(p, reference_line, data_long, x_var)
}
\description{
Helper function to add reference line to juvenile plots
}
\keyword{internal}
