% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{apply_juvenile_theme}
\alias{apply_juvenile_theme}
\title{Helper function to apply theme to juvenile plots}
\usage{
apply_juvenile_theme(p, theme_style, legend_position, rotate_x_labels)
}
\description{
Helper function to apply theme to juvenile plots
}
\keyword{internal}
