% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas_isoparalitorales_peru}
\alias{areas_isoparalitorales_peru}
\title{Áreas isoparalitorales}
\format{
Un data.frame con los siguientes atributos:
\describe{
\item{lon}{Longitud correspondiente del área isoparalitoral}
\item{lat}{Latitud correspondiente del área isoparalitoral}
\item{area}{Código del Área Isoparalitoral}
\item{grad}{Categoría de la latitud cada 0.5 grados (3 - 19.5)}
\item{dc}{Categoría de distancia a la costa (10-200)}
}
}
\source{
Instituto del Mar del Perú (IMARPE)
}
\usage{
areas_isoparalitorales_peru
}
\description{
Conjunto de datos que contiene áreas isoparalitorales de la costa peruana.
Estas áreas representan zonas con características similares a lo largo del litoral.
}
\examples{
data(areas_isoparalitorales_peru)
}
\keyword{datasets}
