% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{build_request_params}
\alias{build_request_params}
\title{Construir parámetros de solicitud}
\usage{
build_request_params(
  start_index,
  batch_size,
  token,
  tipo = 2,
  start_date = "",
  end_date = ""
)
}
\arguments{
\item{start_index}{Índice de inicio para paginación}

\item{batch_size}{Tamaño del lote}

\item{token}{Token de seguridad}

\item{tipo}{Tipo de anuncio}

\item{start_date}{Fecha de inicio (opcional)}

\item{end_date}{Fecha de fin (opcional)}
}
\value{
Lista de parámetros
}
\description{
Crea el conjunto de parámetros para la petición API
}
\keyword{internal}
