% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{buscar_columnas_patron}
\alias{buscar_columnas_patron}
\title{Buscar columnas que contienen un patrón seguido de un número}
\usage{
buscar_columnas_patron(data, patron = "pond_", ordenar = TRUE)
}
\arguments{
\item{data}{Un data frame donde buscar las columnas.}

\item{patron}{El patrón o prefijo a buscar (por defecto "pond_").}

\item{ordenar}{Indica si las columnas deben ordenarse numéricamente (por defecto TRUE).}
}
\value{
Vector de caracteres con los nombres de las columnas que coinciden con el patrón.
}
\description{
Identifica columnas en un dataframe cuyos nombres siguen el patrón
de un prefijo específico seguido de un número.
}
\examples{
# Identificar todas las columnas pond_X en el dataframe

data_calas <- procesar_calas(data_calas = calas_bitacora)
data_faenas <- procesar_faenas(data_faenas = faenas_bitacora)
calas_tallas <- procesar_tallas(data_tallas = tallas_bitacora)

data_tallasfaenas <- merge(x = data_faenas, y = calas_tallas, by = 'codigo_faena')
data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallasfaenas)
datos_final <- agregar_variables(data_total)

tallas_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5","12", "12.5", "13", "13.5", "14", "14.5", "15")
tallas_pond <- ponderar_tallas_df(df = datos_final, tallas_cols = tallas_cols, captura_col = "catch_ANCHOVETA", a= 0.0001, b = 2.984)

cols_pond <- buscar_columnas_patron(data = tallas_pond, patron = "pond_")

# Usar las columnas encontradas para cálculos
tallas_pond[, cols_pond]
}
