% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calas_bitacora}
\alias{calas_bitacora}
\title{Calas de pesca}
\format{
Un data.frame con datos de ejemplo de las calas realizadas por embarcaciones en el mar peruano.
}
\source{
Datos generados de manera aleatoria.
}
\usage{
calas_bitacora
}
\description{
Conjunto de datos que contiene información sobre calas de pesca en el mar peruano.
Las calas son lugares específicos donde se realizan actividades de pesca.
}
\examples{
data(calas_bitacora)
}
\keyword{datasets}
