% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_distancia_grid}
\alias{calcular_distancia_grid}
\title{Calcular distancia de cuadrícula con resolución fija}
\usage{
calcular_distancia_grid(
  lon1,
  lat1,
  lon2,
  lat2,
  resolucion = 0.25,
  unidad = "km"
)
}
\arguments{
\item{lon1}{Longitud del punto base.}

\item{lat1}{Latitud del punto base.}

\item{lon2}{Vector de longitudes de destino.}

\item{lat2}{Vector de latitudes de destino.}

\item{resolucion}{Resolución de la cuadrícula (por defecto 0.25 grados).}

\item{unidad}{Unidad de distancia: "mn" o "km".}
}
\value{
Lista con distancia mínima y su índice.
}
\description{
Calcular distancia de cuadrícula con resolución fija
}
\keyword{internal}
