% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_distancia_haversine_wgs84}
\alias{calcular_distancia_haversine_wgs84}
\title{Calcular distancia usando la fórmula de Haversine con el elipsoide WGS84}
\usage{
calcular_distancia_haversine_wgs84(lon1, lat1, lon2, lat2, unidad)
}
\arguments{
\item{lon1}{Longitud del punto de origen (en grados o radianes).}

\item{lat1}{Latitud del punto de origen (en grados o radianes).}

\item{lon2}{Vector de longitudes destino (en grados o radianes).}

\item{lat2}{Vector de latitudes destino (en grados o radianes).}

\item{unidad}{Unidad de salida: "mn" para millas náuticas, "km" para kilómetros.}
}
\value{
Lista con la distancia mínima y el índice correspondiente.
}
\description{
Calcular distancia usando la fórmula de Haversine con el elipsoide WGS84
}
\keyword{internal}
