% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_distancia_manhattan}
\alias{calcular_distancia_manhattan}
\title{Calcular distancia tipo Manhattan (bloques de ciudad) entre coordenadas}
\usage{
calcular_distancia_manhattan(lon1, lat1, lon2, lat2, unidad)
}
\arguments{
\item{lon1}{Longitud del punto de origen.}

\item{lat1}{Latitud del punto de origen.}

\item{lon2}{Vector de longitudes destino.}

\item{lat2}{Vector de latitudes destino.}

\item{unidad}{Unidad de distancia: "mn" o "km".}
}
\value{
Lista con la distancia mínima y el índice correspondiente.
}
\description{
Calcular distancia tipo Manhattan (bloques de ciudad) entre coordenadas
}
\keyword{internal}
