% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_distancias_vectorizado}
\alias{calcular_distancias_vectorizado}
\title{Calcular distancias entre múltiples puntos y una línea costera (vectorizado)}
\usage{
calcular_distancias_vectorizado(
  lon_punto,
  lat_punto,
  costa_lon,
  costa_lat,
  tipo_distancia,
  ventana,
  unidad
)
}
\arguments{
\item{lon_punto}{Vector de longitudes de puntos a comparar.}

\item{lat_punto}{Vector de latitudes de puntos a comparar.}

\item{costa_lon}{Vector de longitudes de la costa.}

\item{costa_lat}{Vector de latitudes de la costa.}

\item{tipo_distancia}{Tipo de distancia: "haversine", "manhattan" o "grid".}

\item{ventana}{Tamaño de la ventana (en grados) para filtrar la costa. Si es 0, se usa toda la costa.}

\item{unidad}{Unidad de distancia: "mn" o "km".}
}
\value{
Lista con vectores de distancias mínimas e índices correspondientes.
}
\description{
Calcular distancias entre múltiples puntos y una línea costera (vectorizado)
}
