% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{calcular_juveniles}
\alias{calcular_juveniles}
\title{Calcula porcentajes de juveniles para un conjunto de frecuencias}
\usage{
calcular_juveniles(
  frecuencias,
  valores_tallas,
  juvLim = 12,
  a = 0.0012,
  b = 3.1242
)
}
\arguments{
\item{frecuencias}{Vector numérico con las frecuencias por talla.}

\item{valores_tallas}{Vector numérico con los valores de las tallas correspondientes.}

\item{juvLim}{Talla límite para considerar juveniles (por defecto 12 cm).}

\item{a}{Coeficiente de la relación longitud-peso.}

\item{b}{Exponente de la relación longitud-peso.}
}
\value{
Data frame con porcentajes y totales de juveniles.
}
\description{
Función auxiliar que calcula porcentajes de juveniles tanto en número como en peso
a partir de un conjunto de frecuencias de tallas.
}
\examples{
frecuencias <- c(10, 15, 25, 30, 20, 10)
valores_tallas <- c(8, 9, 10, 11, 12, 13)
calcular_juveniles(frecuencias, valores_tallas)
}
