% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_coast_longitude}
\alias{calculate_coast_longitude}
\title{Calculate the approximate longitude of the coast at a specific latitude}
\usage{
calculate_coast_longitude(coastline, latitude)
}
\arguments{
\item{coastline}{A dataframe containing coastline data with at least two columns:
'Lat' (latitude) and 'Long' (longitude) in decimal degrees.}

\item{latitude}{Numeric value representing the latitude in decimal degrees for which
the coast longitude is to be found.}
}
\value{
A numeric value representing the average longitude in decimal degrees of the
coastal points near the specified latitude. If no nearby points are found,
returns -75 as an approximation for the Peruvian coast and displays a warning.
}
\description{
This function finds the approximate longitude of a coastline at a given latitude,
by looking for nearby coastal points and calculating their average longitude.
}
\details{
The function looks for coastal points within a range of 0.1 degrees of the specified
latitude. This threshold may need adjustments depending on the density of points in the dataset.
}
\note{
\itemize{
\item Requires that the input dataframe has the correct structure with columns 'Lat' and 'Long'.
\item The default value of -75 for the Peruvian coast is only a general approximation and
may not be suitable for other regions.
\item This function does not perform interpolations and may give inaccurate results on coasts with
very irregular shapes or deep bays.
\item Works best when coastal data has a uniform distribution of points.
}
}
\keyword{internal}
