% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_distances_vectorized}
\alias{calculate_distances_vectorized}
\title{Calculate distances between multiple points and a coastline (vectorized)}
\usage{
calculate_distances_vectorized(
  lon_point,
  lat_point,
  coast_lon,
  coast_lat,
  distance_type,
  window,
  unit
)
}
\arguments{
\item{lon_point}{Vector of longitudes of points to compare.}

\item{lat_point}{Vector of latitudes of points to compare.}

\item{coast_lon}{Vector of coastline longitudes.}

\item{coast_lat}{Vector of coastline latitudes.}

\item{distance_type}{Type of distance: "haversine", "manhattan" or "grid".}

\item{window}{Window size (in degrees) to filter the coast. If 0, the entire coast is used.}

\item{unit}{Distance unit: "nm" or "km".}
}
\value{
List with vectors of minimum distances and corresponding indices.
}
\description{
Calculate distances between multiple points and a coastline (vectorized)
}
\keyword{internal}
