% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fishery-statistics.R
\name{convert_numbers_to_weight}
\alias{convert_numbers_to_weight}
\title{Convert numbers to weight}
\usage{
convert_numbers_to_weight(data, length_cols, a, b)
}
\arguments{
\item{data}{Data frame with length frequency columns.}

\item{length_cols}{Vector of length column names or numeric values.}

\item{a}{Coefficient of the length-weight relationship.}

\item{b}{Exponent of the length-weight relationship.}
}
\value{
Data frame with original columns plus weight columns (prefixed with "weight_").
}
\description{
Converts numerical length frequencies to weight estimates using length-weight relationship.
}
\examples{
\dontrun{
weight_data <- convert_numbers_to_weight(
  data = frequency_data,
  length_cols = c("8", "8.5", "9", "9.5", "10"),
  a = 0.0012,
  b = 3.1242
)
}

}
