% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crear_poligonos}
\alias{crear_poligonos}
\title{Crear polígonos a partir de coordenadas geográficas y millas}
\usage{
crear_poligonos(
  datos,
  lat_ini_col,
  lat_fin_col,
  lon_ini_col,
  lon_fin_col,
  millas_ini_col = NULL,
  millas_fin_col = NULL
)
}
\arguments{
\item{datos}{Data frame con las columnas necesarias para construir los polígonos.}

\item{lat_ini_col}{Nombre de la columna con la latitud inicial.}

\item{lat_fin_col}{Nombre de la columna con la latitud final.}

\item{lon_ini_col}{Nombre de la columna con la longitud inicial.}

\item{lon_fin_col}{Nombre de la columna con la longitud final.}

\item{millas_ini_col}{(Opcional) Columna con distancia mínima en millas náuticas.}

\item{millas_fin_col}{(Opcional) Columna con distancia máxima en millas náuticas.}
}
\value{
Un objeto \code{sf} con geometría de tipo polígono y atributos asociados.
}
\description{
Crear polígonos a partir de coordenadas geográficas y millas
}
