% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{create_download_dir}
\alias{create_download_dir}
\title{Create download directory}
\usage{
create_download_dir(download_dir, verbose = TRUE)
}
\arguments{
\item{download_dir}{Path to download directory.}

\item{verbose}{Show detailed messages.}
}
\value{
TRUE if directory was created or already exists.
}
\description{
Creates download directory if it doesn't exist.
}
\keyword{internal}
