% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{create_juvenile_base_plot}
\alias{create_juvenile_base_plot}
\title{Helper function to create base plot for juvenile analysis}
\usage{
create_juvenile_base_plot(
  data_long,
  x_var,
  fill_var,
  use_facet_wrap,
  group_by_type
)
}
\description{
Helper function to create base plot for juvenile analysis
}
\keyword{internal}
