% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-helpers-visualization.R
\name{customize_juvenile_axes}
\alias{customize_juvenile_axes}
\title{Helper function to customize axes in juvenile plots}
\usage{
customize_juvenile_axes(p, data_long, x_var, x_date_breaks, y_limits)
}
\description{
Helper function to customize axes in juvenile plots
}
\keyword{internal}
