% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data-imarpe.R
\name{data_juv_imarpe}
\alias{data_juv_imarpe}
\title{Extrae y procesa datos de pesca juvenil y adulta desde una página web de IMARPE}
\usage{
data_juv_imarpe(link)
}
\arguments{
\item{link}{Una cadena de texto que contiene la URL de la página web de IMARPE con la tabla de datos.}
}
\value{
Un \code{data.frame} con las columnas "fecha", "juv_ton", "adu_ton" y "total_ton", donde:
\itemize{
\item "fecha" es la fecha de los datos, convertida a formato \code{Date}.
\item "juv_ton", "adu_ton" y "total_ton" son las cantidades en toneladas de pesca juvenil, adulta y total, respectivamente.
}
}
\description{
La función \code{data_juv_imarpe()} extrae datos de una tabla contenida en la página web de IMARPE (Instituto del Mar del Perú) y los procesa para su análisis.
Esta tabla contiene las cantidades de pesca juvenil, adulta y total (en toneladas) para diferentes fechas. Los datos extraídos son limpiados, convirtiendo
las fechas a un formato estándar y las cantidades a valores numéricos sin espacios.
}
\examples{
link <- "https://www.imarpe.gob.pe/imarpe/archivos/reportes/Reporte_Anchoveta_II_2024.html"
datos <- data_juv_imarpe(link)
print(datos)

}
