% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{distancia_costa}
\alias{distancia_costa}
\title{Distancia a la Costa vectorizado}
\usage{
distancia_costa(
  lon,
  lat,
  linea_costa,
  devolver_indices = FALSE,
  tipo_distancia = "haversine",
  unidad = "mn",
  ventana = 2,
  paralelo = FALSE,
  nucleos = 4
)
}
\arguments{
\item{lon}{Un vector con las longitudes}

\item{lat}{Un vector con las latitudes}

\item{linea_costa}{Un data frame con las coordenadas de la costa, el orden debe ser: lon, lat}
}
\value{
Un vector con las distancias a la costa
}
\description{
Función que permite estimar la distancia entre puntos hacia una lista de puntos
}
\examples{
data_calas = processing_calas(data_calas = calas)
distancia_costa(lon = data_calas$lon_final, lat = data_calas$lat_final, linea_costa = Shoreline_Peru)
}
