% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{download_announcement_file}
\alias{download_announcement_file}
\title{Descargar archivo de anuncio}
\usage{
download_announcement_file(url, file_name, download_dir, verbose = TRUE)
}
\arguments{
\item{url}{URL de descarga}

\item{file_name}{Nombre del archivo}

\item{download_dir}{Directorio de descarga}

\item{verbose}{Mostrar información detallada}
}
\value{
TRUE si se descargó correctamente, FALSE en caso contrario
}
\description{
Descarga un archivo desde la URL proporcionada
}
\keyword{internal}
