% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{download_file}
\alias{download_file}
\title{Download file}
\usage{
download_file(url, file_name, download_dir, verbose = TRUE)
}
\arguments{
\item{url}{Download URL.}

\item{file_name}{Local file name.}

\item{download_dir}{Download directory.}

\item{verbose}{Show progress messages.}
}
\value{
TRUE if download successful, FALSE otherwise.
}
\description{
Internal function to download a file from a URL.
}
\keyword{internal}
