% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encontrar_linea_paralela}
\alias{encontrar_linea_paralela}
\title{Encuentra la línea paralela más cercana a una distancia en millas náuticas}
\usage{
encontrar_linea_paralela(millas, paralelas_costa)
}
\arguments{
\item{millas}{Valor numérico que indica la distancia en millas náuticas que se desea encontrar.}

\item{paralelas_costa}{Lista de data frames con líneas paralelas a diferentes distancias.
Cada elemento debe tener un nombre que indique la distancia (ej. "l5", "l10").}
}
\value{
Una lista con dos elementos:
\itemize{
\item df: El data frame con la línea paralela más cercana a la distancia solicitada.
\item milla_real: El valor numérico de la milla realmente encontrada.
}
}
\description{
Esta función busca dentro de una lista de líneas paralelas a la costa la que mejor
se aproxima a una distancia específica en millas náuticas.
}
\examples{
# Asumiendo que paralelas_costa_peru es una lista de líneas a distintas distancias
resultado <- encontrar_linea_paralela(7, paralelas_costa_peru)
linea <- resultado$df
milla_encontrada <- resultado$milla_real

}
\keyword{internal}
