% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comunicados.R
\name{extract_announcement_data}
\alias{extract_announcement_data}
\title{Extract data from PRODUCE announcements in PDF format}
\usage{
extract_announcement_data(vector_pdf_names)
}
\arguments{
\item{vector_pdf_names}{A character vector of PDF file names or URLs to be processed.}
}
\value{
A data.frame with the extracted information from the announcements. The data.frame includes the following
columns:
\itemize{
\item StartDateTime: Start date and time.
\item EndDateTime: End date and time.
\item StartLatitude: Starting latitude.
\item EndLatitude: Ending latitude.
\item StartLongitude: Starting longitude (if available).
\item EndLongitude: Ending longitude (if available).
\item StartNauticalMiles: Starting nautical miles (if longitude is not available).
\item EndNauticalMiles: Ending nautical miles (if longitude is not available).
\item announcement: Name of the PDF file or announcement.
}
}
\description{
This function processes PDF files containing official announcements and extracts relevant information,
such as dates, times, latitude and longitude coordinates, and nautical miles. If longitude data is not found,
the function will use the nautical mile values instead. The result is a data.frame containing the start and
end dates, coordinates, and other information related to the announcement.
}
\examples{
# Using local files
\dontrun{
pdf_files <- c("announcement1.pdf", "announcement2.pdf")
results <- extract_announcement_data(vector_pdf_names = pdf_files)
}

# Using URLs
pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)

results <- extract_announcement_data(vector_pdf_names = pdf_urls)

print(head(results))

}
