% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{extract_cookies}
\alias{extract_cookies}
\title{Extraer cookies}
\usage{
extract_cookies(response, verbose = TRUE)
}
\arguments{
\item{response}{Respuesta HTTP}

\item{verbose}{Mostrar información detallada}
}
\value{
DataFrame con cookies
}
\description{
Extrae las cookies de una respuesta HTTP
}
\keyword{internal}
