% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{extract_token}
\alias{extract_token}
\title{Extraer token}
\usage{
extract_token(html_content, verbose = TRUE)
}
\arguments{
\item{html_content}{Contenido HTML de la página}

\item{verbose}{Mostrar información detallada}
}
\value{
Token extraído
}
\description{
Extrae el token de seguridad de la página HTML
}
\keyword{internal}
