% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comunicados.R
\name{extrae_data_comunicados}
\alias{extrae_data_comunicados}
\title{Extract data from PRODUCE announcements in PDF format}
\usage{
extrae_data_comunicados(vector_pdf_names)
}
\arguments{
\item{vector_pdf_names}{A character vector of PDF file names or URLs to be processed.}
}
\value{
A data.frame with the extracted information from the announcements. The data.frame includes the following
columns:
\itemize{
\item FechaHoraInicio: Start date and time.
\item FechaHoraFin: End date and time.
\item LatitudInicio: Starting latitude.
\item LatitudFin: Ending latitude.
\item LongitudInicio: Starting longitude (if available).
\item LongitudFin: Ending longitude (if available).
\item MillasNauticasInicio: Starting nautical miles (if longitude is not available).
\item MillasNauticasFin: Ending nautical miles (if longitude is not available).
\item comunicado: Name of the PDF file or announcement.
}
}
\description{
This function processes PDF files containing official announcements and extracts relevant information,
such as dates, times, latitude and longitude coordinates, and nautical miles. If longitude data is not found,
the function will use the nautical mile values instead. The result is a data.frame containing the start and
end dates, coordinates, and other information related to the announcement.
}
\examples{
# Using local files
\dontrun{
pdf_files <- c("comunicado1.pdf", "comunicado2.pdf")
results <- extrae_data_comunicados(vector_pdf_names = pdf_files)
}

# Using URLs
pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)
results <- extrae_data_comunicados(vector_pdf_names = pdf_urls)

}
