% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extraer_puntos_entre_latitudes}
\alias{extraer_puntos_entre_latitudes}
\title{Extrae todos los puntos de una línea paralela entre dos latitudes}
\usage{
extraer_puntos_entre_latitudes(linea, lat_min, lat_max, nombre_lat, nombre_lon)
}
\arguments{
\item{linea}{Data frame que representa una línea paralela a la costa.}

\item{lat_min}{Valor numérico de la latitud mínima (más al sur).}

\item{lat_max}{Valor numérico de la latitud máxima (más al norte).}

\item{nombre_lat}{Nombre de la columna que contiene las latitudes en el data frame.}

\item{nombre_lon}{Nombre de la columna que contiene las longitudes en el data frame.}
}
\value{
Una matriz de puntos con columnas correspondientes a longitud y latitud.
}
\description{
Esta función obtiene todos los puntos de una línea paralela a la costa que se encuentran
entre dos latitudes dadas, incluyendo puntos interpolados en los límites exactos.
}
\details{
La función primero filtra los puntos que están dentro del rango de latitudes.
Si es necesario, interpola puntos adicionales para los límites exactos de
lat_min y lat_max. Si no hay suficientes puntos en el rango, crea un segmento
recto entre los puntos interpolados en los límites.
}
\examples{
# Asumiendo que linea_10_millas es un data frame con columnas 'lat' y 'lon'
puntos <- extraer_puntos_entre_latitudes(paralelas_costa_peru$l10, -14.5, -12.0, "lat", "lon")

}
\keyword{internal}
