% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faenas_bitacora}
\alias{faenas_bitacora}
\title{Faenas de pesca}
\format{
Un data.frame con datos de ejemplo de las faenas realizadas por embarcaciones en el mar peruano.
}
\source{
Datos generados de manera aleatoria.
}
\usage{
faenas_bitacora
}
\description{
Conjunto de datos que contiene información sobre faenas de pesca realizadas en el litoral peruano.
}
\examples{
data(faenas_bitacora)
}
\keyword{datasets}
