% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_parallel_line}
\alias{find_parallel_line}
\title{Find the parallel line closest to a distance in nautical miles}
\usage{
find_parallel_line(miles, coast_parallels)
}
\arguments{
\item{miles}{Numeric value indicating the distance in nautical miles to be found.}

\item{coast_parallels}{List of data frames with parallel lines at different distances.
Each element must have a name indicating the distance (e.g. "l5", "l10").}
}
\value{
A list with two elements:
\itemize{
\item df: The data frame with the parallel line closest to the requested distance.
\item actual_mile: The numeric value of the mile actually found.
}
}
\description{
This function searches within a list of lines parallel to the coast for the one that best
approximates a specific distance in nautical miles.
}
\keyword{internal}
