% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{generate_download_url}
\alias{generate_download_url}
\title{Generar URL de descarga}
\usage{
generate_download_url(file_name)
}
\arguments{
\item{file_name}{Nombre del archivo}
}
\value{
URL completa de descarga
}
\description{
Construye la URL de descarga para un archivo
}
\keyword{internal}
