% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{get_main_page}
\alias{get_main_page}
\title{Obtener p치gina principal}
\usage{
get_main_page(verbose = TRUE)
}
\arguments{
\item{verbose}{Mostrar informaci칩n detallada}
}
\value{
Objeto de respuesta HTTP
}
\description{
Realiza la petici칩n a la p치gina principal y devuelve la respuesta
}
\keyword{internal}
