% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{graficar_poligonos}
\alias{graficar_poligonos}
\title{Graficar polígonos de suspensión pesquera}
\usage{
graficar_poligonos(
  datos,
  costa = Tivy::linea_costa_peru,
  paralelas = Tivy::paralelas_costa_peru,
  tipo = "estatico",
  titulo = "Zonas de suspensión pesquera",
  colores = NULL,
  mostrar_leyenda = FALSE,
  etiquetas = NULL,
  agregar_grid = FALSE,
  capas_base = FALSE,
  minimap = FALSE
)
}
\arguments{
\item{datos}{Lista de polígonos con coordenadas y metadatos (como comunicado).}

\item{costa}{Data frame con la línea de costa a graficar. Default \code{Tivy::linea_costa_peru}.}

\item{paralelas}{Un data frame que contiene las líneas paralelas a la costa. Default \code{Tivy::paralelas_costa_peru}.}

\item{tipo}{Tipo de gráfico a generar: \code{"estatico"} para ggplot2 o \code{"interactivo"} para leaflet.}

\item{titulo}{Título del gráfico.}

\item{colores}{Vector de colores a usar para diferenciar los polígonos (por comunicado u otra categoría).}

\item{mostrar_leyenda}{Lógico. Si \code{TRUE}, se muestra la leyenda. Por defecto es \code{FALSE}.}

\item{etiquetas}{Vector opcional de etiquetas personalizadas para cada polígono. Si no se especifica, se usa el nombre del comunicado.}

\item{agregar_grid}{Lógico. Si \code{TRUE}, agrega una cuadrícula de referencia al gráfico estático.}

\item{capas_base}{Lógico. Si \code{TRUE}, agrega capas base (como satélite o topografía) en el mapa interactivo.}

\item{minimap}{Lógico. Si \code{TRUE}, agrega un minimapa en la esquina del mapa interactivo.}
}
\value{
Un objeto de clase \code{ggplot} si \code{tipo = "estatico"} o un objeto \code{leaflet} si \code{tipo = "interactivo"}.
}
\description{
Esta función permite graficar polígonos de suspensión pesquera utilizando
un mapa base de la costa peruana. Dependiendo del parámetro \code{tipo}, puede
generar un gráfico estático con \code{ggplot2} o un gráfico interactivo con \code{leaflet}.
}
\examples{
# Crear mapa básico
resultados <- extrae_data_comunicados(c("comunicado1.pdf", "comunicado2.pdf"))
graficar_poligonos(datos = resultados, costa = Tivy::linea_costa_peru)

}
