% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{graficar_poligonos_ggplot}
\alias{graficar_poligonos_ggplot}
\title{Graficar polígonos de zonas de suspensión pesquera con ggplot2}
\usage{
graficar_poligonos_ggplot(
  datos,
  costa = Tivy::linea_costa_peru,
  paralelas = Tivy::paralelas_costa_peru,
  titulo = NULL,
  colores = NULL,
  mostrar_leyenda = FALSE,
  etiquetas = NULL,
  agregar_grid = FALSE
)
}
\arguments{
\item{datos}{Un data frame con coordenadas de latitud y longitud de inicio y fin,
que pueden estar en formato de texto (como "12°30'S") o decimales.}

\item{costa}{Un data frame que contiene la línea de costa para visualización.
Debe tener columnas 'Long' y 'Lat'. Default \code{Tivy::linea_costa_peru}.}

\item{paralelas}{Un data frame que contiene las líneas paralelas a la costa. Default \code{Tivy::paralelas_costa_peru}.}

\item{titulo}{Título para el gráfico. Default NULL}

\item{colores}{Vector de colores para los polígonos. Default NULL}

\item{mostrar_leyenda}{Lógico. Si es TRUE, muestra la leyenda. Default FALSE.}

\item{etiquetas}{Vector de textos para etiquetar cada polígono en la leyenda.}

\item{agregar_grid}{Lógico. Si es TRUE, agrega cuadrícula al gráfico. Default FALSE.}
}
\value{
Un objeto ggplot2 que puede ser modificado con capas adicionales.
}
\description{
Crea visualizaciones estáticas de polígonos para zonas de suspensión pesquera
utilizando ggplot2. Devuelve un objeto ggplot2 que puede ser modificado posteriormente.
}
\examples{
# Crear gráfico básico

pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)
results <- extrae_data_comunicados(vector_pdf_names = pdf_urls)

g <- graficar_poligonos_ggplot(datos = results, costa = Tivy::linea_costa_peru)

# Personalizar posteriormente
g +
  ggplot2::xlim(-80, -70) +
  ggplot2::theme_minimal() +
  ggplot2::labs(title = "Título personalizado")

}
