% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{info_columnas_patron}
\alias{info_columnas_patron}
\title{Obtener nombres y posiciones de columnas que coinciden con un patrón}
\usage{
info_columnas_patron(data, patron = "pond_", ordenar = TRUE)
}
\arguments{
\item{data}{Un data frame donde buscar las columnas.}

\item{patron}{El patrón o prefijo a buscar (por defecto "pond_").}

\item{ordenar}{Indica si los resultados deben ordenarse numéricamente (por defecto TRUE).}
}
\value{
Lista con dos elementos: "posiciones" y "nombres" de las columnas coincidentes.
}
\description{
Devuelve tanto los nombres como las posiciones de columnas en un dataframe
que siguen el patrón especificado.
}
\examples{
# Obtener tanto posiciones como nombres

data_calas <- procesar_calas(data_calas = calas_bitacora)
data_faenas <- procesar_faenas(data_faenas = faenas_bitacora)
calas_tallas <- procesar_tallas(data_tallas = tallas_bitacora)

data_tallasfaenas <- merge(x = data_faenas, y = calas_tallas, by = 'codigo_faena')
data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallasfaenas)
datos_final <- agregar_variables(data_total)

tallas_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5","12", "12.5", "13", "13.5", "14", "14.5", "15")
tallas_pond <- ponderar_tallas_df(df = datos_final, tallas_cols = tallas_cols, captura_col = "catch_ANCHOVETA", a= 0.0001, b = 2.984)

resultado <- info_columnas_patron(tallas_pond, "pond_")

posiciones <- resultado$posiciones
print(posiciones)
nombres <- resultado$nombres
print(nombres)

}
