% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolar_punto}
\alias{interpolar_punto}
\title{Interpola un punto para una latitud específica en una línea paralela}
\usage{
interpolar_punto(linea, latitud, nombre_lat, nombre_lon)
}
\arguments{
\item{linea}{Data frame que representa una línea paralela a la costa.}

\item{latitud}{Valor numérico de la latitud para la cual se desea encontrar la longitud.}

\item{nombre_lat}{Nombre de la columna que contiene las latitudes en el data frame.}

\item{nombre_lon}{Nombre de la columna que contiene las longitudes en el data frame.}
}
\value{
Un vector con dos elementos:
\itemize{
\item lon: Longitud interpolada para la latitud dada.
\item lat: La misma latitud proporcionada como entrada.
}
}
\description{
Esta función calcula la longitud correspondiente a una latitud específica en una línea
paralela a la costa, interpolando entre los puntos existentes cuando sea necesario.
}
\details{
La función ordena la línea por latitud, busca los puntos que encierran la latitud
deseada y realiza una interpolación lineal. En caso de no encontrar puntos adecuados
para interpolar, devuelve el punto más cercano.
}
\keyword{internal}
