% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{interpolate_point}
\alias{interpolate_point}
\title{Interpolate point on line}
\usage{
interpolate_point(line, latitude, lat_name, lon_name)
}
\arguments{
\item{line}{Data frame representing a parallel line.}

\item{latitude}{Target latitude.}

\item{lat_name}{Name of latitude column.}

\item{lon_name}{Name of longitude column.}
}
\value{
Vector with interpolated longitude and latitude.
}
\description{
Internal function to interpolate a point for a specific latitude on a parallel line.
}
\keyword{internal}
