% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{juvenile_percentage}
\alias{juvenile_percentage}
\title{Percentage of juveniles}
\usage{
juvenile_percentage(frequency, length, juvLim = 12, silence_warnings = FALSE)
}
\arguments{
\item{frequency}{A numeric vector with the frequency of sampled length.}

\item{length}{Vector of length corresponding to the frequencies.}

\item{juvLim}{length limit (default 12 cm) to classify as juvenile.}

\item{silence_warnings}{Logical. If TRUE, warning messages are suppressed (default = FALSE).}
}
\value{
Percentage of juveniles in the sample.
}
\description{
Estimates the percentage of individuals considered juveniles in a sample
according to an established length limit.
}
\examples{

frequency <- c(0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2)

length <- seq(5, 20, by = 1)

perc <- juvenile_percentage(frequency, length, juvLim = 12)

print(perc)
}
