% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{juveniles_por_grupo}
\alias{juveniles_por_grupo}
\title{Versión mejorada utilizando dplyr}
\usage{
juveniles_por_grupo(
  data,
  group_cols,
  cols_tallas,
  juvLim = 12,
  a = 0.0012,
  b = 3.1242,
  remove_empty = TRUE
)
}
\arguments{
\item{data}{Data frame con datos de frecuencias de tallas.}

\item{group_cols}{Vector de nombres de columnas para agrupar los datos.}

\item{juvLim}{Talla límite para considerar juveniles (por defecto 12 cm).}

\item{a}{Coeficiente de la relación longitud-peso.}

\item{b}{Exponente de la relación longitud-peso.}

\item{tallas_cols}{Vector de nombres de columnas que contienen las frecuencias de tallas.}
}
\value{
Data frame con porcentajes de juveniles por grupos, tanto en número como en peso.
}
\description{
Esta versión utiliza el paquete dplyr para un enfoque más moderno
y efectivo del cálculo de juveniles por grupo, aprovechando la función porc_juveniles existente.
}
\examples{
data_calas <- procesar_calas(data_calas = calas_bitacora)
data_faenas <- procesar_faenas(data_faenas = faenas_bitacora)
calas_tallas <- procesar_tallas(data_tallas = tallas_bitacora)

data_tallasfaenas <- merge(x = data_faenas, y = calas_tallas, by = 'codigo_faena')
data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallasfaenas)

datos_final <- agregar_variables(data_total)

tallas_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5","12", "12.5", "13", "13.5", "14", "14.5", "15")
resultado <- ponderar_tallas_df(df = datos_final, tallas_cols = tallas_columnas, captura_col = "catch_ANCHOVETA", a= 0.0001, b = 2.984)

resultado$fecha_unica <- convertir_a_fecha(resultado$fecha_inicio_cala, tipo = "date")

resultado_juveniles <- juveniles_por_grupo(data = resultado, group_cols = c("fecha_unica", "dc_cat"), cols_tallas = tallas_cols, juvLim = 12, a = 0.0012, b = 3.1242))
}
