% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{length_weight}
\alias{length_weight}
\title{Length - Weight Relationship}
\usage{
length_weight(length, a, b)
}
\arguments{
\item{length}{A numeric vector containing the lengths of the individuals.}

\item{a}{Numeric value of the coefficient of the length-weight relationship.}

\item{b}{Numeric value of the exponent of the length-weight relationship.}
}
\value{
A numeric vector with the estimated weights.
}
\description{
This function estimates the weight of an individual from its length, using
the general formula of the length-weight relationship: \eqn{W = a \cdot L^b}, where
\eqn{W} is the weight, \eqn{L} the length (length), \eqn{a} and \eqn{b} the specific parameters.
}
\examples{

lengths <- seq(5, 20, by = 0.5)
a <- 0.0001
b <- 2.984

weights <- length_weight(length = lengths, a = a, b = b)

print(weights)
}
