% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linea_costa_peru}
\alias{linea_costa_peru}
\title{Línea costera de Perú}
\format{
Un objeto espacial (sf o sp) con los siguientes atributos:
\describe{
\item{geometry}{Geometría lineal que representa la costa peruana}
\item{...}{Otros atributos relevantes de la costa}
}
}
\source{
Indicar la fuente de los datos: "Institución que proporcionó los datos"
}
\usage{
linea_costa_peru
}
\description{
Conjunto de datos que contiene la línea costera de Perú representada como un objeto espacial.
}
\examples{
data(linea_costa_peru)
}
\keyword{datasets}
