% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{make_api_request}
\alias{make_api_request}
\title{Realizar petición API}
\usage{
make_api_request(url, params, cookies, main_url)
}
\arguments{
\item{url}{URL del endpoint}

\item{params}{Parámetros de la petición}

\item{cookies}{Cookies para la petición}

\item{main_url}{URL de referencia}
}
\value{
Objeto de respuesta HTTP
}
\description{
Realiza la petición POST al API con los parámetros proporcionados
}
\keyword{internal}
