% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{min_range}
\alias{min_range}
\alias{max_range}
\title{Máximo y Mínimo de la talla}
\usage{
min_range(frecuencia, tallas)

max_range(frecuencia, tallas)
}
\arguments{
\item{frecuencia}{Un vector con las frecuencias de las tallas muestreadas en la captura}

\item{tallas}{Un vector numérico que representa las tallas muestreadas}
}
\value{
La talla mínima muestreada
}
\description{
Funciones para determinar el máximo y mínimo de la talla
}
\examples{
frecuencia <- c(0, 0, 0, 0, 0, 0, 0, 0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2, 0, 0, 0, 0, 0, 0, 0, 0)
tallas <- seq(from = 5, to = 20, by = 0.5)
min_range(frecuencia = frecuencia, tallas = tallas)
max_range(frecuencia = frecuencia, tallas = tallas)
}
