% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{max_range}
\alias{max_range}
\title{Talla máxima observada con frecuencia positiva}
\usage{
max_range(frecuencia, tallas)
}
\arguments{
\item{frecuencia}{Un vector numérico con las frecuencias de tallas.}

\item{tallas}{Vector de tallas correspondiente.}
}
\value{
Valor máximo de talla con frecuencia mayor que cero.
}
\description{
Talla máxima observada con frecuencia positiva
}
\examples{
max_range(frecuencia = c(0,0,1,2,3), tallas = c(5,6,7,8,9))
}
