% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_faenas_tallas.R
\name{merge_tallas_faenas_calas}
\alias{merge_tallas_faenas_calas}
\title{Mergeando datos de faenas, tallas y calas}
\usage{
merge_tallas_faenas_calas(data_calas, data_tallas_faenas)
}
\arguments{
\item{data_calas}{Un data frame obtenido de processing_calas()}

\item{data_tallas_faenas}{Un data frame con los datos de tallas de las faenas a procesar}
}
\value{
Un data frame con los datos consolidados de faenas, tallas y calas
}
\description{
merge_tallas_faenas_calas() Función para juntar los archivos de faenas, tallas y calas
}
\examples{
data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallas_faenas)
}
