% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{number_to_weight}
\alias{number_to_weight}
\title{Conversion from number of individuals to weight}
\usage{
number_to_weight(data, length, a, b)
}
\arguments{
\item{data}{Data frame where columns with names equal to length contain frequencies.}

\item{length}{Vector of length (which must match the column names of \code{data}).}

\item{a}{Numeric value of the coefficient of the length-weight relationship.}

\item{b}{Numeric value of the exponent of the length-weight relationship.}
}
\value{
Data frame with the same dimensions but expressed in weight.
}
\description{
Converts numerical length frequencies in a data.frame to weight estimates,
using the length-weight relationship.
}
\examples{

data(calas_bitacora, faenas_bitacora, tallas_bitacora)

# Process data
data_hauls <- process_hauls(data_hauls = calas_bitacora)
data_fishing_trips <- process_fishing_trips(data_fishing_trips = faenas_bitacora)
hauls_length <- process_length(data_length = tallas_bitacora)

# Integrate data
data_length_fishing_trips <- merge(x = data_fishing_trips, y = hauls_length, by = 'fishing_trip_code')
data_total <- merge_length_fishing_trips_hauls(data_hauls = data_hauls,
                                       data_length_fishing_trips = data_length_fishing_trips)
final_data <- add_variables(data_total)

# Define length columns
length_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5",
                 "12", "12.5", "13", "13.5", "14", "14.5", "15")

# Weight length
results <- weight_length_df(df = final_data,
                               length_cols = length_cols,
                               catch_col = "catch_ANCHOVETA",
                               a = 0.0001,
                               b = 2.984)

number_to_weight(data = results, length = paste0('pond_', length_cols), a = 0.0012, b = 3.1242)
}
