% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{numero_a_peso}
\alias{numero_a_peso}
\title{Conversión de número de individuos a peso}
\usage{
numero_a_peso(data, tallas, a, b)
}
\arguments{
\item{data}{Data frame donde las columnas con nombres iguales a las tallas contienen frecuencias.}

\item{tallas}{Vector de tallas (que deben coincidir con los nombres de columnas del \code{data}).}

\item{a}{Valor numérico del coeficiente de la relación longitud-peso.}

\item{b}{Valor numérico del exponente de la relación longitud-peso.}
}
\value{
Data frame con las mismas dimensiones pero expresadas en peso.
}
\description{
Convierte frecuencias numéricas de tallas en un data.frame a estimaciones de peso,
usando la relación talla-peso.
}
\examples{
data <- data.frame(id = 1:2, `8` = c(3,2), `9` = c(5,4), `10` = c(2,3))
tallas <- c(8,9,10)
numero_a_peso(data, tallas, a = 0.0012, b = 2.984)
}
