% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peru_coast_parallels}
\alias{peru_coast_parallels}
\title{Lines parallel to the Peruvian coast}
\format{
A list of data.frames, each representing a line parallel to the Peruvian coast.
\describe{
Each data frame contains:
\item{lon}{Longitude corresponding to the isoparalittoral area}
\item{lat}{Latitude corresponding to the isoparalittoral area}
\item{dc}{Distance to coast category (10-200)}
}
}
\source{
Indicate the data source
}
\usage{
peru_coast_parallels
}
\description{
Dataset containing lines parallel to the Peruvian coast at different distances.
These lines are useful for spatial analyses related to fishery management.
}
\examples{
data(peru_coast_parallels)
}
\keyword{datasets}
