% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peru_coastline}
\alias{peru_coastline}
\title{Peruvian coastline}
\format{
A data.frame with the following fields:
\describe{
\item{Long}{Longitude of the Peruvian coastline}
\item{Lat}{Latitude of the Peruvian coastline}
}
}
\source{
Marine Institute of Peru (IMARPE)
}
\usage{
peru_coastline
}
\description{
Dataset containing the coastline of Peru represented as a spatial object.
}
\examples{
data(peru_coastline)
}
\keyword{datasets}
