% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{peru_isoparalitoral_areas}
\alias{peru_isoparalitoral_areas}
\title{Isoparalittoral areas}
\format{
A data.frame with the following attributes:
\describe{
\item{lon}{Longitude corresponding to the isoparalittoral area}
\item{lat}{Latitude corresponding to the isoparalittoral area}
\item{area}{Isoparalittoral Area code}
\item{grad}{Latitude category every 0.5 degrees (3 - 19.5)}
\item{dc}{Distance to coast category (10-200)}
}
}
\source{
Marine Institute of Peru (IMARPE)
}
\usage{
peru_isoparalitoral_areas
}
\description{
Dataset containing isoparalittoral areas of the Peruvian coast.
These areas represent zones with similar characteristics along the littoral.
}
\examples{
data(peru_isoparalitoral_areas)
}
\keyword{datasets}
