% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Generate interactive plot of polygons with leaflet}
\usage{
plot_interactive(
  polygons,
  coastline,
  title,
  colors,
  show_legend = TRUE,
  labels = NULL,
  base_layers = TRUE,
  minimap = TRUE
)
}
\arguments{
\item{polygons}{List of polygons. Each must have fields such as \code{coords}, \code{announcement}, dates and coordinates.}

\item{coastline}{Data frame with the coastline (columns \code{Long} and \code{Lat}).}

\item{title}{Title to display at the top of the map.}

\item{colors}{Vector of colors. If \code{NULL}, they are automatically assigned with \code{RColorBrewer::brewer.pal}.}

\item{show_legend}{Logical. If \code{TRUE}, the layers control (legend) is displayed.}

\item{labels}{Optional vector of names to display in the legend and map labels.}

\item{base_layers}{Logical. If \code{TRUE}, includes base layers such as satellite and ocean maps.}

\item{minimap}{Logical. If \code{TRUE}, displays a minimap in the lower right corner.}
}
\value{
A \code{leaflet} object with the interactive map.
}
\description{
Creates an interactive map using \code{leaflet}, showing polygons with popup information.
}
\keyword{internal}
