% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_polygons_ggplot}
\alias{plot_polygons_ggplot}
\title{Plot fishing suspension zone polygons with ggplot2}
\usage{
plot_polygons_ggplot(
  data,
  coastline = peru_coastline,
  parallels = peru_coast_parallels,
  title = NULL,
  colors = NULL,
  show_legend = FALSE,
  labels = NULL,
  add_grid = FALSE
)
}
\arguments{
\item{data}{A data frame with latitude and longitude coordinates for start and end points,
which can be in text format (like "12°30'S") or decimal degrees.}

\item{coastline}{A data frame containing the coastline for visualization.
Must have columns 'Long' and 'Lat'. Default \code{Tivy::peru_coastline}.}

\item{parallels}{A data frame containing lines parallel to the coast. Default \code{Tivy::peru_coast_parallels}.}

\item{title}{Title for the plot. Default NULL}

\item{colors}{Vector of colors for the polygons. Default NULL}

\item{show_legend}{Logical. If TRUE, shows the legend. Default FALSE.}

\item{labels}{Vector of texts to label each polygon in the legend.}

\item{add_grid}{Logical. If TRUE, adds a grid to the plot. Default FALSE.}
}
\value{
A ggplot2 object that can be modified with additional layers.
}
\description{
Creates static visualizations of polygons for fishing suspension zones
using ggplot2. Returns a ggplot2 object that can be modified later.
}
\examples{
# Create basic plot

pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)
results <- extract_announcement_data(vector_pdf_names = pdf_urls)

g <- plot_polygons_ggplot(data = results, coastline = Tivy::peru_coastline)

# Customize later
g +
  ggplot2::xlim(-80, -70) +
  ggplot2::theme_minimal() +
  ggplot2::labs(title = "Custom title")

}
