% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ponderacion}
\alias{ponderacion}
\title{Ponderación de la talla la captura}
\usage{
ponderacion(frecuencia, captura, tallas, a, b)
}
\arguments{
\item{frecuencia}{Un vector con las frecuencias de las tallas muestreadas en la captura}

\item{captura}{Un valor numérico que representa la captura}

\item{tallas}{Un vector numérico que representa las tallas muestreadas}

\item{a}{Un valor numérico de la relación longitud - peso}

\item{b}{Un valor numérico de la relación longitud - peso}
}
\value{
Un vector con las tallas ponderadas a la captura
}
\description{
Función para ponderar las tallas en función de la captura.
}
\examples{
frecuencia_tallas <- c(0, 0, 0, 0, 0, 0, 0, 0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2, 0, 0, 0, 0, 0, 0, 0, 0)
captura <- 100
tallas <- seq(from = 5, to = 20, by = 0.5)
a <- 0.0001
b <- 2.984
tallas_ponderadas <- ponderacion(frecuencia = frecuencia_tallas, captura = captura, tallas = tallas, a = a, b = b)
}
