% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ponderar_tallas_df}
\alias{ponderar_tallas_df}
\title{Ponderación de tallas en un data frame}
\usage{
ponderar_tallas_df(
  df,
  tallas_cols,
  captura_col,
  a,
  b,
  paralelo = FALSE,
  num_cores = NULL,
  tam_bloque = 10000,
  silenciar_warnings = TRUE
)
}
\arguments{
\item{df}{Un data frame que contiene las columnas de tallas y captura.}

\item{tallas_cols}{Un vector de caracteres con los nombres de las columnas que representan las tallas.}

\item{captura_col}{Nombre de la columna que contiene los valores de captura.}

\item{a}{Valor numérico del coeficiente de la relación longitud-peso.}

\item{b}{Valor numérico del exponente de la relación longitud-peso.}

\item{paralelo}{Booleano indicando si se debe utilizar procesamiento paralelo.}

\item{num_cores}{Número de núcleos a utilizar (si paralelo=TRUE).}

\item{tam_bloque}{Tamaño de los bloques para procesamiento (si paralelo=TRUE).}
}
\value{
Un data frame con las columnas originales y las columnas de tallas ponderadas.
Las columnas ponderadas tendrán el prefijo "pond_" seguido del nombre original.
}
\description{
Esta función realiza la ponderación de frecuencias de tallas según la captura total
utilizando la relación longitud-peso. Permite procesamiento paralelo para conjuntos
de datos grandes.
}
\examples{

data(calas_bitacora)
data(calas_bitacora)
data(tallas_bitacora)

data_calas <- procesar_calas(data_calas = calas_bitacora)
data_faenas <- procesar_faenas(data_faenas = calas_bitacora)
calas_tallas <- procesar_tallas(data_tallas = tallas_bitacora)

data_tallasfaenas <- merge(x = data_faenas, y = calas_tallas, by = 'codigo_faena')

data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallasfaenas)

tallas_columnas <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5","12", "12.5", "13", "13.5", "14", "14.5", "15")

# Procesamiento secuencial
resultado <- ponderar_tallas_df(df = data_total, tallas_cols = tallas_columnas, captura_col = "catch_ANCHOVETA", a= 0.0001, b = 2.984)

print(head(resultado))

# Procesamiento paralelo para conjuntos grandes
resultado_paralelo <- ponderar_tallas_df(
  df = data_total, tallas_cols = tallas_columnas, captura_col = "catch_ANCHOVETA", a = 0.0001, b = 2.984, paralelo = TRUE
)

print(head(resultado_paralelo))

}
