% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{porc_juveniles}
\alias{porc_juveniles}
\title{Porcentaje de Juveniles}
\usage{
porc_juveniles(frecuencia, tallas, juvLim = 12)
}
\arguments{
\item{frecuencia}{Un vector con las frecuencias de las tallas muestreadas en la captura}

\item{tallas}{Un vector numérico que representa las tallas muestreadas}

\item{juvLim}{Un valor numérico que representa la talla límite para considerar un individuo como juvenil}
}
\value{
El porcentaje de juveniles estimado
}
\description{
Función para estimar el porcentaje de juveniles
}
\examples{
frecuencia <- c(0, 0, 0, 0, 0, 0, 0, 0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2, 0, 0, 0, 0, 0, 0, 0, 0)
tallas <- seq(from = 5, to = 20, by = 0.5)
juveniles <- porc_juveniles(frecuencia = frecuencia, tallas = tallas, juvLim = 12)
}
