% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_polygons}
\alias{prepare_polygons}
\title{Prepare polygons from coordinate data}
\usage{
prepare_polygons(
  data,
  coastline,
  coast_parallels = NULL,
  column_names = c("lat", "lon", "dc")
)
}
\arguments{
\item{data}{A data frame with coordinates.}

\item{coastline}{A data frame with the coastline.}

\item{coast_parallels}{List of data frames with lines parallel to the coast at different distances.
Each element of the list must have a name indicating the distance (e.g. "l5", "l10").}

\item{column_names}{A vector with column names in the coast_parallels data frames.
Must contain: c("lat", "lon", "dc") where "dc" is the distance to the coast.}
}
\value{
A list of polygons for visualization.
}
\description{
Helper function to process data and prepare polygons using preexisting parallel lines.
The east and west edges of the polygon will follow the shape of the lines parallel
to the coast at the specified nautical miles, while the north and south edges will be
straight lines at the given latitudes.
}
\keyword{internal}
