% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{procesar_bloque}
\alias{procesar_bloque}
\title{Procesa un bloque de filas para la ponderación de tallas}
\usage{
procesar_bloque(df_bloque, tallas_cols, captura_col, a, b)
}
\arguments{
\item{df_bloque}{Data frame a procesar}

\item{tallas_cols}{Nombres de las columnas de tallas}

\item{captura_col}{Nombre de la columna de captura}

\item{a}{Coeficiente de la relación longitud-peso}

\item{b}{Exponente de la relación longitud-peso}
}
\value{
Data frame con columnas ponderadas añadidas con el prefijo "pond_"
}
\description{
Función auxiliar que procesa un conjunto de filas para calcular las tallas ponderadas.
Puede ser utilizada directamente para procesar subconjuntos de datos.
}
\keyword{internal}
