% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_block}
\alias{process_block}
\title{Process a block of rows for length weighting}
\usage{
process_block(df, length_cols, catch_col, a, b, silence_warnings = FALSE)
}
\arguments{
\item{df}{Data frame to process}

\item{length_cols}{Names of the length columns}

\item{catch_col}{Name of the catch column}

\item{a}{Coefficient of the length-weight relationship}

\item{b}{Exponent of the length-weight relationship}

\item{silence_warnings}{Whether to suppress warnings}
}
\value{
Data frame with weighted columns added with the prefix "pond_"
}
\description{
Helper function that processes a set of rows to calculate weighted length.
Can be used directly to process data subsets.
}
\keyword{internal}
