% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-faenas.R
\name{process_fishing_trips}
\alias{process_fishing_trips}
\title{Processing fishing trip data from PRODUCE sitrapesca files}
\usage{
process_fishing_trips(data_fishing_trips, format = "xlsx")
}
\arguments{
\item{data_fishing_trips}{Data frame with raw fishing trip data.}

\item{format}{Format of the input file: "xlsx" (default) or "csv".}
}
\value{
A data frame with the following standardized columns:
\itemize{
\item fishing_trip_code: Unique identifier of the fishing trip
\item vessel: Vessel name
\item owner: Vessel owner
\item id_vessel: Vessel registration number
\item start_date: Start date and time of the fishing trip (optional, only in XLSX format)
\item end_date: End date and time of the fishing trip (optional, only in XLSX format)
}
}
\description{
Processes fishing trip data from PRODUCE logbooks,
in CSV or XLSX format. Returns a clean and standardized data frame.
}
\examples{

data(faenas_bitacora)

fishing_trips = process_fishing_trips(data_fishing_trips = faenas_bitacora, format = "xlsx")

print(head(fishing_trips))
}
