% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{process_json_response}
\alias{process_json_response}
\title{Process JSON response}
\usage{
process_json_response(content_text, verbose = TRUE)
}
\arguments{
\item{content_text}{Response text.}

\item{verbose}{Show detailed information.}
}
\value{
Data frame with announcements or NULL if error.
}
\description{
Processes JSON response from API and extracts announcements.
}
\keyword{internal}
