% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-tallas.R
\name{process_length}
\alias{process_length}
\title{Processing length data from hauls, PRODUCE sitrapesca file}
\usage{
process_length(data_length, format = "xlsx")
}
\arguments{
\item{data_length}{Data frame with length data.}

\item{format}{Input format: "xlsx" (default) or "csv".}
}
\value{
A data frame with length by haul (wide format).
}
\description{
Processes length data from PRODUCE logbooks in CSV or XLSX format.
Cleans the relevant columns and transforms the long format to wide format (rows by haul, columns by length).
}
\examples{

process_length(data_length = tallas_bitacora, format = "xlsx")

}
