% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-calas.R
\name{processing_calas}
\alias{processing_calas}
\alias{processing_calas_2}
\title{Procesamiento de datos de calas}
\usage{
processing_calas(data_calas)

processing_calas_2(data_calas)
}
\arguments{
\item{data_calas}{Un data frame con los datos de calas a procesar}
}
\value{
Un data frame con los datos procesados
}
\description{
processing_calas() Funci칩n para procesar datos de calas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato xlsx

processing_calas_2() Funci칩n para procesar datos de calas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato csv
}
\examples{
data_calas <- processing_calas(data_calas = calas)
}
