% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-faenas.R
\name{processing_faenas}
\alias{processing_faenas}
\alias{processing_faenas_2}
\title{Procesamiento de datos de faenas}
\usage{
processing_faenas(data_faenas)

processing_faenas_2(data_faenas)
}
\arguments{
\item{data_faenas}{Un data frame con los datos de faenas a procesar}
}
\value{
Un data frame con los datos procesados
}
\description{
processing_faenas() Funci칩n para procesar datos de faenas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato xlsx

processing_faenas_2() Funci칩n para procesar datos de faenas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato csv
}
\examples{
data_faenas <- processing_faenas(data_faenas = faenas)
}
