% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing-tallas.R
\name{processing_tallas}
\alias{processing_tallas}
\alias{processing_tallas_2}
\title{Procesamiento de datos de tallas de las calas}
\usage{
processing_tallas(data_tallas)

processing_tallas_2(data_tallas)
}
\arguments{
\item{data_tallas}{Un data frame con los datos de tallas de las calas a procesar}
}
\value{
Un data frame con los datos procesados
}
\description{
processing_tallas() Funci칩n para procesar datos de tallas de las calas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato xlsx

processing_tallas_2() Funci칩n para procesar datos de tallas de las calas pesqueras proveniente de las bit치coras de PRODUCE. Usa formato csv
}
\examples{
data_tallas_calas <- processing_tallas(data_tallas = faenas)
}
