% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{puntos_tierra}
\alias{puntos_tierra}
\title{Puntos en tierra}
\usage{
puntos_tierra(x_punto, y_punto, linea_costa, paralelo = FALSE, nucleos = 4)
}
\arguments{
\item{linea_costa}{Un data frame con las coordenadas de la costa, el orden debe ser: lon, lat}

\item{lon}{Un vector con las longitudes}

\item{lat}{Un vector con las latitudes}
}
\value{
Un vector con las distancias a la costa
}
\description{
Función que permite estimar los puntos que están a la derecha o izquierda de la línea de costa
}
\examples{
data_calas = processing_calas(data_calas = calas)
puntos_tierra(x_punto = data_calas$lon_final, y_punto = data_calas$lat_final, linea_costa = Shoreline_Peru)
}
