% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{search_pattern_columns}
\alias{search_pattern_columns}
\title{Search for columns that contain a pattern followed by a number}
\usage{
search_pattern_columns(data, pattern = "pond_", sort = TRUE)
}
\arguments{
\item{data}{A data frame to search for columns.}

\item{pattern}{The pattern or prefix to search for (default "pond_").}

\item{sort}{Indicates if columns should be sorted numerically (default TRUE).}
}
\value{
Character vector with the names of columns that match the pattern.
}
\description{
Identifies columns in a dataframe whose names follow the pattern
of a specific prefix followed by a number.
}
\examples{

data(calas_bitacora, faenas_bitacora, tallas_bitacora)

# Identify all pond_X columns in the dataframe
data_hauls <- process_hauls(data_hauls = calas_bitacora)
data_fishing_trips <- process_fishing_trips(data_fishing_trips = faenas_bitacora)
hauls_length <- process_length(data_length = tallas_bitacora)

data_length_trips <- merge(x = data_fishing_trips, y = hauls_length, by = 'fishing_trip_code')
data_total <- merge_length_fishing_trips_hauls(data_hauls = data_hauls, data_length_fishing_trips = data_length_trips)
final_data <- add_variables(data_total)

length_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5","12", "12.5", "13", "13.5", "14", "14.5", "15")
length_weighted <- weight_length_df(df = final_data, length_cols = length_cols, catch_col = "catch_ANCHOVETA", a = 0.0001, b = 2.984)

pond_cols <- search_pattern_columns(data = length_weighted, pattern = "pond_")

# Use the found columns for calculations
length_weighted[, pond_cols]
}
