% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{talla_peso}
\alias{talla_peso}
\title{Relación Talla - Peso}
\usage{
talla_peso(talla, a, b)
}
\arguments{
\item{talla}{Un vector numérico que contiene las tallas de los individuos.}

\item{a}{Valor numérico del coeficiente de la relación longitud-peso.}

\item{b}{Valor numérico del exponente de la relación longitud-peso.}
}
\value{
Un vector numérico con los pesos estimados.
}
\description{
Esta función estima el peso de un individuo a partir de su talla, utilizando
la fórmula general de la relación longitud-peso: \eqn{W = a \cdot L^b}, donde
\eqn{W} es el peso, \eqn{L} la longitud (talla), \eqn{a} y \eqn{b} los parámetros específicos.
}
\examples{
tallas <- seq(5, 20, by = 0.5)
a <- 0.0001
b <- 2.984
pesos <- talla_peso(talla = tallas, a = a, b = b)
}
