% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tallas_bitacora}
\alias{tallas_bitacora}
\title{Datos de tallas muestreadas provenientes de las calas realizadas por las embarcaciones en el mar peruano.}
\format{
Un data.frame con datos de ejemplo de tallas muestreadas provenientes de las calas realizadas por embarcaciones en el mar peruano.
}
\source{
Indicar la fuente de los datos
}
\usage{
tallas_bitacora
}
\description{
Conjunto de datos que contiene información sobre tallas de especies marinas
capturadas en el litoral peruano.
}
\examples{
data(tallas_bitacora)
}
\keyword{datasets}
