% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Tivy-package}
\alias{Tivy}
\alias{Tivy-package}
\title{Tivy: Toolkit for Investigation and Visualization of Young Anchovies}
\description{
Specialized toolkit for processing biological and fisheries data from Peru's anchovy (Engraulis ringens) fishery. Provides functions to analyze fishing logbooks, calculate biological indicators (length-weight relationships, juvenile percentages), generate spatial fishing indicators, and visualize regulatory measures from Peru's Ministry of Production. Features automated data processing from multiple file formats, coordinate validation, spatial analysis of fishing zones, and tools for analyzing fishing closure announcements and regulatory compliance. Includes built-in datasets of Peruvian coastal coordinates and parallel lines for analyzing fishing activities within regulatory zones.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HansTtito/Tivy}
  \item Report bugs at \url{https://github.com/HansTtito/Tivy/issues}
}

}
\author{
\strong{Maintainer}: Hans Ttito \email{kvttitos@gmail.com} (\href{https://orcid.org/0000-0003-3732-9419}{ORCID})

}
\keyword{internal}
