% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{weight_length_df}
\alias{weight_length_df}
\title{Weighting length in a data frame}
\usage{
weight_length_df(
  df,
  length_cols,
  catch_col,
  a,
  b,
  parallel = FALSE,
  num_cores = NULL,
  block_size = 10000,
  silence_warnings = TRUE
)
}
\arguments{
\item{df}{A data frame containing the length columns and catch.}

\item{length_cols}{A character vector with the names of the columns that represent the length.}

\item{catch_col}{Name of the column that contains the catch values.}

\item{a}{Numeric value of the coefficient of the length-weight relationship.}

\item{b}{Numeric value of the exponent of the length-weight relationship.}

\item{parallel}{Boolean indicating whether to use parallel processing.}

\item{num_cores}{Number of cores to use (if parallel=TRUE).}

\item{block_size}{Size of the blocks for processing (if parallel=TRUE).}

\item{silence_warnings}{Logical. If TRUE, warning messages are suppressed (default = TRUE).}
}
\value{
A data frame with the original columns and the weighted length columns.
The weighted columns will have the prefix "pond_" followed by the original name.
}
\description{
This function performs the weighting of length frequencies according to the total catch
using the length-weight relationship. It allows parallel processing for large datasets.
}
\examples{

data(calas_bitacora, faenas_bitacora, tallas_bitacora)

data_hauls <- process_hauls(data_hauls = calas_bitacora)
data_fishing_trips <- process_fishing_trips(data_fishing_trips = faenas_bitacora)
hauls_length <- process_length(data_length = tallas_bitacora)

data_length_fishing_trips <- merge(
   x = data_fishing_trips, 
   y = hauls_length, 
   by = 'fishing_trip_code'
)

data_total <- merge_length_fishing_trips_hauls(
   data_hauls = data_hauls, 
   data_length_fishing_trips = data_length_fishing_trips
)

length_columns <- c("8", "8.5", "9", "9.5", "10", "10.5",
                    "11", "11.5","12", "12.5", "13", "13.5",
                    "14", "14.5", "15")

# Sequential processing
results <- weight_length_df(
   df = data_total, 
   length_cols = length_columns, 
   catch_col = "catch_ANCHOVETA", 
   a= 0.0001,
   b = 2.984
)

print(head(results))

# Parallel processing for large datasets
parallel_results <- weight_length_df(
   df = data_total, 
   length_cols = length_columns, 
   catch_col = "catch_ANCHOVETA", 
   a = 0.0001,
   b = 2.984,
   parallel = TRUE
)

print(head(parallel_results))

}
