% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{weighting}
\alias{weighting}
\title{Weighting length according to total catch}
\usage{
weighting(frequency, catch, length, a, b, silence_warnings = FALSE)
}
\arguments{
\item{frequency}{A numeric vector with the frequency of length observed in the sampling.}

\item{catch}{Numeric value with the total catch (in kg or tons, as appropriate).}

\item{length}{A numeric vector with the length corresponding to the frequencies.}

\item{a}{Numeric value of the coefficient of the length-weight relationship.}

\item{b}{Numeric value of the exponent of the length-weight relationship.}

\item{silence_warnings}{Logical. If TRUE, warning messages are suppressed (default = FALSE).}
}
\value{
A numeric vector with the weighted frequencies.
}
\description{
Calculates a weighting of the sampled length based on the total recorded catch.
This allows scaling the observed frequency to the total catch.
}
\examples{

frequency <- c(0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2)

length <- seq(5, 20, by = 1)

catch <- 1000
a <- 0.0001
b <- 2.984

weighted <- weighting(frequency, catch, length, a, b)

print(head(weighted))
}
