% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-variables.R
\name{add_variables}
\alias{add_variables}
\title{Adds variables for juveniles, sample length, distance to coast, and distance category}
\usage{
add_variables(
  data,
  JuvLim = 12,
  distance_type = "haversine",
  window = 0.5,
  unit = "nm",
  coastline = NULL,
  suppress_warnings = TRUE
)
}
\arguments{
\item{data}{A data frame that must contain latitude (\code{lat_initial}) and longitude (\code{lon_initial}) coordinates,
as well as columns with individual length.}

\item{JuvLim}{Length threshold to consider juveniles (default = 12). If the length is below this value,
the individual is considered juvenile.}

\item{distance_type}{Type of distance calculation to the coast (e.g., "haversine"), default is "haversine".}

\item{window}{Window parameter to smooth the coastline, default is 0.5.}

\item{unit}{Distance unit used in the calculation ("nm", "km", etc.), default is "nm".}

\item{coastline}{\code{data.frame} with coastline coordinates. Must have columns named \code{Long} and \code{Lat}. If \code{NULL} (default), uses internal dataset \code{Tivy::peru_coastline}.}

\item{suppress_warnings}{Logical. If TRUE (default), warnings are suppressed; otherwise, they are shown.}
}
\value{
A data frame with the following new variables:
\itemize{
\item \code{juv}: Proportion of juveniles in each row.
\item \code{sample}: Total number of individuals in the sample.
\item \code{dc}: Distance to the coast based on provided latitude and longitude.
\item \code{dc_cat}: Categorical variable for distance to the coast (e.g., "05-15 nm", "15-30 nm").
}
}
\description{
This function adds new variables to a dataset, including the proportion of juveniles,
the total number of individuals in the sample, the distance to the coast, and the distance category.
}
\examples{

data(calas_bitacora, faenas_bitacora, tallas_bitacora)

# Process data
data_hauls <- process_hauls(data_hauls = calas_bitacora)
data_fishing_trips <- process_fishing_trips(data_fishing_trips = faenas_bitacora)
hauls_length <- process_length(data_length = tallas_bitacora)

# Merge hauls, length and fishing trips
data_length_trips <- merge(
 x = data_fishing_trips, 
 y = hauls_length, 
 by = 'fishing_trip_code'
)
data_total <- merge_length_fishing_trips_hauls(
 data_hauls = data_hauls, 
 data_length_fishing_trips = data_length_trips
)

# Apply function
results <- add_variables(data = data_total)

print(results)
}
