% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-variables.R
\name{agregar_variables}
\alias{agregar_variables}
\title{Agrega variables de juveniles, muestra, distancia a la costa y su categoría}
\usage{
agregar_variables(
  data,
  JuvLim = 12,
  tipo_distancia = "haversine",
  ventana = 0.5,
  unidad = "mn"
)
}
\arguments{
\item{data}{Data frame con tallas y coordenadas}

\item{JuvLim}{Límite de talla para considerar juveniles (default = 12)}

\item{tipo_distancia}{Tipo de cálculo de distancia ("haversine", etc.)}

\item{ventana}{Ventana para suavizar línea de costa (default = 0.5)}

\item{unidad}{Unidad de distancia ("mn", "km", etc.)}
}
\value{
Data frame con nuevas variables: juv, muestra, dc y dc_cat
}
\description{
Agrega variables de juveniles, muestra, distancia a la costa y su categoría
}
