% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-variables.R
\name{agregar_variables}
\alias{agregar_variables}
\title{agregando variables a la base de datos}
\usage{
agregar_variables(data)
}
\arguments{
\item{data}{Un data frame obtenido de merge_tallas_faenas_calas()}
}
\value{
Un data frame con los datos consolidados de faenas, tallas y calas
}
\description{
agregar_variables() Función que permite agregar variables como la distancia a la costa, la cantidad de juveniles y la cantidad de muestras
}
\examples{
data_total <- agregar_variables(data = data_total)
}
