% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{build_request_params}
\alias{build_request_params}
\title{Build request parameters}
\usage{
build_request_params(
  start_index,
  batch_size,
  token,
  tipo = 2,
  start_date = "",
  end_date = ""
)
}
\arguments{
\item{start_index}{Start index for pagination.}

\item{batch_size}{Batch size.}

\item{token}{Security token.}

\item{tipo}{Announcement type.}

\item{start_date}{Start date (optional).}

\item{end_date}{End date (optional).}
}
\value{
List of parameters.
}
\description{
Creates parameter set for API request.
}
\keyword{internal}
