% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calas_bitacora}
\alias{calas_bitacora}
\title{Fishing hauls}
\format{
A data.frame with sample data of hauls made by vessels in the Peruvian sea.
}
\source{
Randomly generated data.
}
\usage{
calas_bitacora
}
\description{
Dataset containing information about fishing hauls in the Peruvian sea.
Hauls are specific locations where fishing activities are conducted.
}
\examples{
data(calas_bitacora)
}
\keyword{datasets}
