% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{calculate_coast_longitude}
\alias{calculate_coast_longitude}
\title{Calculate coast longitude}
\usage{
calculate_coast_longitude(coastline, latitude)
}
\arguments{
\item{coastline}{Coastline data frame.}

\item{latitude}{Target latitude.}
}
\value{
Approximate coast longitude.
}
\description{
Internal function to calculate approximate longitude of coast at a given latitude.
}
\keyword{internal}
