% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fishery-statistics.R
\name{calculate_fish_weight}
\alias{calculate_fish_weight}
\title{Calculate fish weight from length}
\usage{
calculate_fish_weight(length, a, b)
}
\arguments{
\item{length}{Numeric vector of fish lengths.}

\item{a}{Coefficient of the length-weight relationship.}

\item{b}{Exponent of the length-weight relationship.}
}
\value{
Numeric vector of estimated weights.
}
\description{
Estimates individual fish weight from length using the length-weight relationship: W = a * L^b
}
\examples{
lengths <- seq(8, 20, by = 0.5)
weights <- calculate_fish_weight(length = lengths, a = 0.0001, b = 2.984)

}
