% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_haversine_distance_wgs84}
\alias{calculate_haversine_distance_wgs84}
\title{Calculate distance using the Haversine formula with the WGS84 ellipsoid}
\usage{
calculate_haversine_distance_wgs84(lon1, lat1, lon2, lat2, unit)
}
\arguments{
\item{lon1}{Longitude of the origin point (in degrees or radians).}

\item{lat1}{Latitude of the origin point (in degrees or radians).}

\item{lon2}{Vector of destination longitudes (in degrees or radians).}

\item{lat2}{Vector of destination latitudes (in degrees or radians).}

\item{unit}{Output unit: "nm" for nautical miles, "km" for kilometers.}
}
\value{
List with the minimum distance and the corresponding index.
}
\description{
Calculate distance using the Haversine formula with the WGS84 ellipsoid
}
\keyword{internal}
