% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fishery-statistics.R
\name{calculate_juvenile_percentage}
\alias{calculate_juvenile_percentage}
\title{Calculate juvenile percentage}
\usage{
calculate_juvenile_percentage(
  frequency,
  length,
  juvenile_limit = 12,
  silence_warnings = FALSE
)
}
\arguments{
\item{frequency}{Numeric vector of length frequencies.}

\item{length}{Numeric vector of corresponding lengths.}

\item{juvenile_limit}{Length threshold for juvenile classification.}

\item{silence_warnings}{Logical. Suppress warnings.}
}
\value{
Percentage of juveniles in the sample.
}
\description{
Calculates the percentage of individuals considered juveniles based on a length threshold.
}
\examples{
freq <- c(0, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 6, 0)
lengths <- seq(8, 20, by = 1)
juv_pct <- calculate_juvenile_percentage(freq, lengths, juvenile_limit = 12)

}
