% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_manhattan_distance}
\alias{calculate_manhattan_distance}
\title{Calculate Manhattan distance (city blocks) between coordinates}
\usage{
calculate_manhattan_distance(lon1, lat1, lon2, lat2, unit)
}
\arguments{
\item{lon1}{Longitude of the origin point.}

\item{lat1}{Latitude of the origin point.}

\item{lon2}{Vector of destination longitudes.}

\item{lat2}{Vector of destination latitudes.}

\item{unit}{Distance unit: "nm" or "km".}
}
\value{
List with the minimum distance and the corresponding index.
}
\description{
Calculate Manhattan distance (city blocks) between coordinates
}
\keyword{internal}
