% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{check_required_packages}
\alias{check_required_packages}
\title{Verificar paquetes requeridos}
\usage{
check_required_packages()
}
\value{
TRUE si los paquetes están disponibles
}
\description{
Comprueba si están instalados los paquetes necesarios
}
\keyword{internal}
