% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-time-functions.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Converts a vector of dates in different formats to a standard date format}
\usage{
convert_to_date(date_vector, type = "datetime")
}
\arguments{
\item{date_vector}{A character vector containing dates in various formats.}

\item{type}{Type of object to return: "date" for Date, "datetime" for POSIXct (default).}
}
\value{
A vector of \code{Date} or \code{POSIXct} objects according to the type parameter, or \code{NA} if the date cannot be converted.
}
\description{
The \code{convert_to_date()} function takes a vector of dates in various formats and converts them to a standard date format in R.
The function tries to parse each date using a series of predefined formats and returns the first valid date found for each entry.
If a date cannot be interpreted by any of the formats, it is assigned as \code{NA}.
}
\examples{

dates <- c("2025-04-10", "10/04/2025", "April 10, 2025")

converted_dates <- convert_to_date(dates)

print(converted_dates)

}
