% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{create_download_dir}
\alias{create_download_dir}
\title{Crear directorio de descarga}
\usage{
create_download_dir(download_dir, verbose = TRUE)
}
\arguments{
\item{download_dir}{Ruta del directorio de descarga}

\item{verbose}{Mostrar información detallada}
}
\value{
TRUE si se creó el directorio o ya existía
}
\description{
Crea el directorio para descargar archivos si no existe
}
\keyword{internal}
