% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dashboard_juveniles}
\alias{dashboard_juveniles}
\title{Crear dashboard visual para el análisis de juveniles}
\usage{
dashboard_juveniles(
  datos_juveniles,
  var_x = NULL,
  limite_juv = NULL,
  paleta = NULL,
  ordenar_comparacion = FALSE
)
}
\arguments{
\item{datos_juveniles}{Data frame generado por \code{juveniles_por_grupo}, que
contiene los porcentajes de juveniles por grupo.}

\item{var_x}{Nombre de la variable a utilizar en el eje X de los gráficos
(por ejemplo, "fecha_unica" para visualización temporal).}

\item{limite_juv}{(Opcional) Valor umbral para el porcentaje legal de
juveniles. Si se proporciona, se añade como línea de referencia en los gráficos.}

\item{paleta}{(Opcional) Vector de colores personalizado para los gráficos.}

\item{ordenar_comparacion}{Lógico. Indica si el gráfico de comparación de
porcentajes debe ser ordenado de mayor a menor (por defecto es FALSE).}
}
\value{
Una lista de objetos \code{ggplot} que representan distintos aspectos del
análisis de juveniles, lista para ser integrada en un dashboard.
}
\description{
Genera un conjunto de visualizaciones en formato dashboard que resumen
los resultados obtenidos a partir de la función \code{juveniles_por_grupo}.
Este dashboard permite explorar la proporción de juveniles según diferentes
criterios agrupados, incluyendo comparaciones por número y peso,
así como tendencias temporales.
}
\examples{
# Ejemplo de uso:
data_calas <- procesar_calas(data_calas = calas_bitacora)
data_faenas <- procesar_faenas(data_faenas = faenas_bitacora)
calas_tallas <- procesar_tallas(data_tallas = tallas_bitacora)

data_tallasfaenas <- merge(x = data_faenas, y = calas_tallas, by = "codigo_faena")
data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallasfaenas)

datos_final <- agregar_variables(data_total)
tallas_cols <- c("8", "8.5", "9", "9.5", "10", "10.5", "11", "11.5", "12",
                 "12.5", "13", "13.5", "14", "14.5", "15")
resultado <- ponderar_tallas_df(df = datos_final, tallas_cols = tallas_cols,
                                captura_col = "catch_ANCHOVETA", a = 0.0001, b = 2.984)
resultado$fecha_unica <- convertir_a_fecha(resultado$fecha_inicio_cala, tipo = "date")
resultado_juveniles <- juveniles_por_grupo(data = resultado,
                                           group_cols = c("fecha_unica"),
                                           cols_tallas = tallas_cols)

dashboard <- dashboard_juveniles(
  datos_juveniles = resultado_juveniles,
  var_x = "fecha_unica"
)
}
