% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{dms_a_decimal}
\alias{dms_a_decimal}
\title{Convertir latitud y longitud a grados decimales}
\usage{
dms_a_decimal(coordenadas, hemisferio = "S")
}
\arguments{
\item{coordenadas}{Un vector con las longitudes o latitudes a convertir}
}
\value{
Un vector con las longitudes o latitudes convertidas a decimales
}
\description{
Función que permite convertir longitud o latitude en grados a decimales.
}
\examples{
dms_a_decimal(calas$Longitud.Fin)
}
