% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{extract_cookies}
\alias{extract_cookies}
\title{Extract cookies}
\usage{
extract_cookies(response, verbose = TRUE)
}
\arguments{
\item{response}{HTTP response.}

\item{verbose}{Show detailed information.}
}
\value{
Data frame with cookies.
}
\description{
Extracts cookies from HTTP response.
}
\keyword{internal}
