% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{extract_points_between_latitudes}
\alias{extract_points_between_latitudes}
\title{Extract points between latitudes}
\usage{
extract_points_between_latitudes(line, lat_min, lat_max, lat_name, lon_name)
}
\arguments{
\item{line}{Data frame representing a line parallel to the coast.}

\item{lat_min}{Minimum latitude (southernmost).}

\item{lat_max}{Maximum latitude (northernmost).}

\item{lat_name}{Name of latitude column.}

\item{lon_name}{Name of longitude column.}
}
\value{
Matrix of points with longitude and latitude columns.
}
\description{
Extracts all points of a parallel line between two latitudes.
}
\keyword{internal}
