% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{extract_token}
\alias{extract_token}
\title{Extract token}
\usage{
extract_token(html_content, verbose = TRUE)
}
\arguments{
\item{html_content}{HTML content of the page.}

\item{verbose}{Show detailed information.}
}
\value{
Extracted token.
}
\description{
Extracts security token from HTML page.
}
\keyword{internal}
