% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comunicados.R
\name{extrae_data_comunicados}
\alias{extrae_data_comunicados}
\title{Extrae datos de los comunicados en formato PDF}
\usage{
extrae_data_comunicados(vector_pdf_names)
}
\arguments{
\item{vector_pdf_names}{Un vector de nombres de archivo PDF a procesar.}
}
\value{
Un data.frame con la información extraída de los comunicados. El data.frame incluye las siguientes
columnas:
\itemize{
\item FechaHoraInicio: Fecha y hora de inicio.
\item FechaHoraFin: Fecha y hora de fin.
\item LatitudInicio: Latitud de inicio.
\item LatitudFin: Latitud de fin.
\item LongitudInicio: Longitud de inicio (si está disponible).
\item LongitudFin: Longitud de fin (si está disponible).
\item MillasNauticasInicio: Millas náuticas de inicio (si no hay longitudes).
\item MillasNauticasFin: Millas náuticas de fin (si no hay longitudes).
\item comunicado: Nombre del archivo del comunicado.
}
}
\description{
Esta función procesa archivos PDF que contienen comunicados y extrae información relevante, como fechas,
horas, coordenadas de latitud y longitud, y millas náuticas. Si no se encuentran longitudes, la función
procesará las millas náuticas. El resultado es un data.frame con las fechas de inicio y fin, coordenadas
y otra información relacionada con el comunicado.
}
\examples{
# Suponiendo que tengas una lista de archivos PDF
pdf_files <- c("comunicado1.pdf", "comunicado2.pdf")
resultados <- extrae_data_comunicados(pdf_files)
head(resultados)

}
