% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{extraer_valores_tallas}
\alias{extraer_valores_tallas}
\title{Extraer valores numéricos desde nombres de columnas}
\usage{
extraer_valores_tallas(nombres_columnas)
}
\arguments{
\item{nombres_columnas}{Vector de caracteres con los nombres de columnas.}
}
\value{
Vector numérico con los valores de tallas extraídos.
}
\description{
Esta función extrae todos los valores numéricos (incluidos decimales) desde un vector
de nombres de columnas, eliminando cualquier carácter no numérico.
}
\examples{
nombres <- c("pond_10.5", "talla_11", "12.5mm", "T14", "13-erróneo")
extraer_valores_tallas(nombres)
# Resultado: 10.5 11.0 12.5 14.0 13.0
}
