% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{faenas_bitacora}
\alias{faenas_bitacora}
\title{Fishing trips}
\format{
A data.frame with sample data of fishing trips made by vessels in the Peruvian sea.
}
\source{
Randomly generated data.
}
\usage{
faenas_bitacora
}
\description{
Dataset containing information about fishing trips conducted along the Peruvian littoral.
}
\examples{
data(faenas_bitacora)
}
\keyword{datasets}
