% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faenas_bitacora}
\alias{faenas_bitacora}
\title{Faenas de pesca}
\format{
Un data.frame o objeto espacial con los siguientes atributos:
\describe{
\item{id}{Identificador único de la faena}
\item{fecha}{Fecha de la faena}
\item{embarcacion}{Identificador o nombre de la embarcación}
\item{...}{Otros atributos relevantes de las faenas de pesca}
}
}
\source{
Indicar la fuente de los datos
}
\usage{
faenas_bitacora
}
\description{
Conjunto de datos que contiene información sobre faenas de pesca realizadas en el litoral peruano.
}
\examples{
data(faenas_bitacora)
# Código de ejemplo para explorar los datos
}
\keyword{datasets}
