% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{fetch_announcements_batch}
\alias{fetch_announcements_batch}
\title{Fetch announcements batch}
\usage{
fetch_announcements_batch(
  start_index,
  batch_size,
  token,
  cookies,
  main_url,
  tipo = 2,
  start_date = "",
  end_date = "",
  verbose = TRUE
)
}
\arguments{
\item{start_index}{Start index for pagination.}

\item{batch_size}{Batch size.}

\item{token}{Security token.}

\item{cookies}{Cookies for request.}

\item{main_url}{Reference URL.}

\item{tipo}{Announcement type.}

\item{start_date}{Start date (optional).}

\item{end_date}{End date (optional).}

\item{verbose}{Show detailed information.}
}
\value{
Data frame with announcements or NULL if error.
}
\description{
Retrieves batch of announcements using token.
}
\keyword{internal}
