% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{find_parallel_line}
\alias{find_parallel_line}
\title{Find parallel line by distance}
\usage{
find_parallel_line(miles, coast_parallels)
}
\arguments{
\item{miles}{Distance in nautical miles.}

\item{coast_parallels}{List of parallel line data frames.}
}
\value{
List with data frame and actual mile value.
}
\description{
Internal function to find the parallel line closest to a specified distance.
}
\keyword{internal}
