% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{generate_download_url}
\alias{generate_download_url}
\title{Generate download URL}
\usage{
generate_download_url(
  file_name,
  base_url =
    "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/"
)
}
\arguments{
\item{file_name}{Name of the file to download.}

\item{base_url}{Base URL for downloads.}
}
\value{
Complete download URL.
}
\description{
Internal function to construct download URLs for files.
}
\keyword{internal}
