% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{get_main_page}
\alias{get_main_page}
\title{Get main page}
\usage{
get_main_page(url, verbose = TRUE)
}
\arguments{
\item{url}{Main page URL.}

\item{verbose}{Show detailed information.}
}
\value{
HTTP response object.
}
\description{
Makes request to main page and returns HTTP response.
}
\keyword{internal}
