% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_point}
\alias{interpolate_point}
\title{Interpolates a point for a specific latitude on a parallel line}
\usage{
interpolate_point(line, latitude, lat_name, lon_name)
}
\arguments{
\item{line}{Data frame representing a line parallel to the coast.}

\item{latitude}{Numeric value of the latitude for which the longitude is to be found.}

\item{lat_name}{Name of the column containing latitudes in the data frame.}

\item{lon_name}{Name of the column containing longitudes in the data frame.}
}
\value{
A vector with two elements:
\itemize{
\item lon: Interpolated longitude for the given latitude.
\item lat: The same latitude provided as input.
}
}
\description{
This function calculates the longitude corresponding to a specific latitude on a line
parallel to the coast, interpolating between existing points when necessary.
}
\details{
The function sorts the line by latitude, finds the points that enclose the desired
latitude and performs a linear interpolation. If suitable points for interpolation
are not found, it returns the closest point.
}
\keyword{internal}
