% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{talla_peso}
\alias{talla_peso}
\title{Relación Talla - Peso}
\usage{
talla_peso(talla, a, b)
}
\arguments{
\item{talla}{Un vector con las tallas}

\item{a}{Un valor numérico de la relación longitud - peso}

\item{b}{Un valor numérico de la relación longitud - peso}
}
\value{
Un vector con los pesos estimados
}
\description{
Función que permite estimar el Peso en función de la talla
}
\examples{
tallas <- seq(from = 5, to = 20, by = 0.5)
a <- 0.0001
b <- 2.984
pesos <- talla_peso(talla = tallas, a = a, b = b)
}
