% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linea_costa_peru}
\alias{linea_costa_peru}
\title{Línea costera de Perú}
\format{
Un data.frame con los siguientes campos:
\describe{
\item{Long}{Longitud de la línea de costa del Perú}
\item{Lat}{Latitud de la línea de costa del Perú}
}
}
\source{
Instituto del Mar del Perú (IMARPE)
}
\usage{
linea_costa_peru
}
\description{
Conjunto de datos que contiene la línea costera de Perú representada como un objeto espacial.
}
\examples{
data(linea_costa_peru)
}
\keyword{datasets}
