% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{make_api_request}
\alias{make_api_request}
\title{Make API request}
\usage{
make_api_request(url, params, cookies, main_url)
}
\arguments{
\item{url}{Endpoint URL.}

\item{params}{Request parameters.}

\item{cookies}{Cookies for request.}

\item{main_url}{Reference URL.}
}
\value{
HTTP response object.
}
\description{
Makes POST request to API with provided parameters.
}
\keyword{internal}
