% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{max_range}
\alias{max_range}
\title{Maximum observed length with positive frequency}
\usage{
max_range(frequency, length)
}
\arguments{
\item{frequency}{A numeric vector with the frequencies of length.}

\item{length}{Corresponding length vector.}
}
\value{
Maximum length value with frequency greater than zero.
}
\description{
Maximum observed length with positive frequency
}
\examples{
max_range(frequency = c(0,0,1,2,3), length = c(5,6,7,8,9))
}
