% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-faenas-tallas.R
\name{merge_length_fishing_trips_hauls}
\alias{merge_length_fishing_trips_hauls}
\title{Merging data from fishing trips, length and hauls}
\usage{
merge_length_fishing_trips_hauls(data_hauls, data_length_fishing_trips)
}
\arguments{
\item{data_hauls}{A data frame processed with \code{process_hauls()}.}

\item{data_length_fishing_trips}{A data frame with length data by fishing trip and haul.}
}
\value{
A data frame with the consolidated data from fishing trips, length and hauls.
}
\description{
\code{merge_length_fishing_trips_hauls()} is a function to join data from fishing trips, length and hauls, combining catches by species,
length ranges (minimum and maximum) and the spatial and temporal information of each haul.
}
\examples{

data(calas_bitacora, faenas_bitacora, tallas_bitacora)

data_hauls <- process_hauls(data_hauls = calas_bitacora, format = "xlsx")
data_fishing_trips <- process_fishing_trips(data_fishing_trips = faenas_bitacora, format = "xlsx")
data_length <- process_length(data_length = tallas_bitacora, format = "xlsx")

data_length_fishing_trips <- merge(x = data_length, y = data_fishing_trips, by = "fishing_trip_code", all = TRUE)

data_total <- merge_length_fishing_trips_hauls(data_hauls = data_hauls, data_length_fishing_trips = data_length_fishing_trips)

print(head(data_total))

}
