% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_faenas_tallas.R
\name{merge_tallas_faenas_calas}
\alias{merge_tallas_faenas_calas}
\title{Mergeando datos de faenas, tallas y calas}
\usage{
merge_tallas_faenas_calas(data_calas, data_tallas_faenas)
}
\arguments{
\item{data_calas}{Un data frame procesado con \code{procesar_calas()}.}

\item{data_tallas_faenas}{Un data frame con los datos de tallas por faena y cala.}
}
\value{
Un data frame con los datos consolidados de faenas, tallas y calas.
}
\description{
\code{merge_tallas_faenas_calas()} es una función para unir los datos de faenas, tallas y calas, combinando las capturas por especie,
los rangos de tallas (mínimo y máximo) y la información espacial y temporal de cada cala.
}
\examples{

data_calas <- procesar_calas(data_calas = calas_bitacora, formato = "xlsx")
data_faenas <- procesar_faenas(data_faenas = faenas_bitacora, formato = "xlsx")
data_tallas <- procesar_tallas(data_tallas = tallas_bitacora, formato = "xlsx")

data_tallas_faenas <- merge(x = data_tallas, y = data_faenas, by = "codigo_faena", all = TRUE)

data_total <- merge_tallas_faenas_calas(data_calas = data_calas, data_tallas_faenas = data_tallas_faenas)

print(head(data_total))

}
