% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{min_range}
\alias{min_range}
\title{Minimum observed length with positive frequency}
\usage{
min_range(frequency, length)
}
\arguments{
\item{frequency}{A numeric vector with the frequencies of length.}

\item{length}{Corresponding length vector.}
}
\value{
Minimum length value with frequency greater than zero.
}
\description{
Minimum observed length with positive frequency
}
\examples{
min_range(frequency = c(0,0,1,2,3), length = c(5,6,7,8,9))
}
