% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{numero_a_peso}
\alias{numero_a_peso}
\title{numero_a_peso}
\usage{
numero_a_peso(data, tallas, a, b)
}
\arguments{
\item{data}{Un data frame con las frecuencias de las tallas en número}

\item{tallas}{Un vector numérico que representa las tallas muestreadas}

\item{a}{Un valor numérico de la relación longitud - peso}

\item{b}{Un valor numérico de la relación longitud - peso}
}
\value{
El porcentaje de juveniles estimado
}
\description{
Función para convertir las tallas en número a peso
}
\examples{
data_tallas <- processing_tallas(data_tallas = tallas)
tallas <- seq(from = 8, to = 15, by = 0.5)
a <- 0.0012
b <- 2.984
data_peso <- numero_a_peso(data = data_tallas, tallas = tallas, a = a, b = b)
}
