% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paralelas_costa_peru}
\alias{paralelas_costa_peru}
\title{Líneas paralelas a la costa peruana}
\format{
Una lista de data.frames, cada uno representa una línea paralela a la costa peruana.
\describe{
Cada data frame contiene:
\item{lon}{Longitud correspondiente del área isoparalitoral}
\item{lat}{Latitud correspondiente del área isoparalitoral}
\item{dc}{Categoría de distancia a la costa (10-200)}
}
}
\source{
Indicar la fuente de los datos
}
\usage{
paralelas_costa_peru
}
\description{
Conjunto de datos que contiene líneas paralelas a la costa peruana a diferentes distancias.
Estas líneas son útiles para análisis espaciales relacionados con la gestión pesquera.
}
\examples{
data(paralelas_costa_peru)
}
\keyword{datasets}
