% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_polygons}
\alias{plot_polygons}
\title{Plot fishing suspension polygons}
\usage{
plot_polygons(
  data,
  coastline = peru_coastline,
  parallels = peru_coast_parallels,
  type = "static",
  title = "Fishing suspension zones",
  colors = NULL,
  show_legend = FALSE,
  labels = NULL,
  add_grid = FALSE,
  base_layers = FALSE,
  minimap = FALSE
)
}
\arguments{
\item{data}{List of polygons with coordinates and metadata (such as announcement).}

\item{coastline}{Data frame with the coastline to plot. Default \code{Tivy::peru_coastline}.}

\item{parallels}{A data frame containing lines parallel to the coast. Default \code{Tivy::peru_coast_parallels}.}

\item{type}{Type of plot to generate: \code{"static"} for ggplot2 or \code{"interactive"} for leaflet.}

\item{title}{Plot title.}

\item{colors}{Vector of colors to use to differentiate the polygons (by announcement or other category).}

\item{show_legend}{Logical. If \code{TRUE}, the legend is shown. Default is \code{FALSE}.}

\item{labels}{Optional vector of custom labels for each polygon. If not specified, the announcement name is used.}

\item{add_grid}{Logical. If \code{TRUE}, adds a reference grid to the static plot.}

\item{base_layers}{Logical. If \code{TRUE}, adds base layers (such as satellite or topography) to the interactive map.}

\item{minimap}{Logical. If \code{TRUE}, adds a minimap in the corner of the interactive map.}
}
\value{
An object of class \code{ggplot} if \code{type = "static"} or a \code{leaflet} object if \code{type = "interactive"}.
}
\description{
This function allows plotting fishing suspension polygons using
a base map of the Peruvian coast. Depending on the \code{type} parameter, it can
generate a static plot with \code{ggplot2} or an interactive plot with \code{leaflet}.
}
\examples{
# Create basic map

pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)

results <- extract_announcement_data(vector_pdf_names = pdf_urls)

plot_polygons(data = results, coastline = Tivy::peru_coastline)

}
