% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_polygons_leaflet}
\alias{plot_polygons_leaflet}
\title{Plot fishing suspension zone polygons with leaflet}
\usage{
plot_polygons_leaflet(
  data,
  coastline = NULL,
  parallels = NULL,
  title = NULL,
  colors = NULL,
  show_legend = FALSE,
  labels = NULL,
  base_layers = FALSE,
  minimap = FALSE
)
}
\arguments{
\item{data}{A data frame with latitude and longitude coordinates for start and end points,
which can be in text format (like "12°30'S") or decimal degrees.}

\item{coastline}{\code{data.frame} with coastline coordinates. Must have columns named \code{Long} and \code{Lat}. If \code{NULL} (default), uses internal dataset \code{Tivy::peru_coastline}.}

\item{parallels}{A list containing data frames of lines parallel to the coast. If \code{NULL} (default), uses internal dataset \code{Tivy::peru_coast_parallels}.}

\item{title}{Title for the plot. Default NULL}

\item{colors}{Vector of colors for the polygons. Default NULL}

\item{show_legend}{Logical. If TRUE, shows the legend. Default FALSE.}

\item{labels}{Vector of texts to label each polygon in the legend.}

\item{base_layers}{Logical. If TRUE, adds multiple base maps. Default FALSE.}

\item{minimap}{Logical. If TRUE, adds a minimap to the interactive map. Default FALSE}
}
\value{
A leaflet object that can be modified with additional layers.
}
\description{
Creates interactive visualizations of polygons for fishing suspension zones
using leaflet. Returns a leaflet object that can be modified later.
}
\examples{
# Create basic map

pdf_urls <- c(
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1542_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1478_comunicado1.pdf",
  "https://consultasenlinea.produce.gob.pe/produce/descarga/comunicados/dgsfs/1468_comunicado1.pdf"
)

results <- extract_announcement_data(vector_pdf_names = pdf_urls)

m <- plot_polygons_leaflet(data = results)

# Customize later
m |>
  leaflet::addMarkers(lng = -77.1, lat = -12.0, popup = "Lima") |>
  leaflet::addCircleMarkers(lng = -76.3, lat = -13.4, radius = 5, color = "red")

}
