% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ponderacion}
\alias{ponderacion}
\title{Ponderación de tallas según captura total}
\usage{
ponderacion(frecuencia, captura, tallas, a, b)
}
\arguments{
\item{frecuencia}{Un vector numérico con la frecuencia de tallas observadas en el muestreo.}

\item{captura}{Valor numérico con la captura total (en kg o toneladas, según el caso).}

\item{tallas}{Un vector numérico con las tallas correspondientes a las frecuencias.}

\item{a}{Valor numérico del coeficiente de la relación longitud-peso.}

\item{b}{Valor numérico del exponente de la relación longitud-peso.}
}
\value{
Un vector numérico con las frecuencias ponderadas.
}
\description{
Calcula una ponderación de las tallas muestreadas en función de la captura total
registrada. Esto permite escalar la frecuencia observada a la captura total.
}
\examples{
frecuencia <- c(0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2)
tallas <- seq(5, 20, by = 1)
captura <- 1000
a <- 0.0001
b <- 2.984
ponderadas <- ponderacion(frecuencia, captura, tallas, a, b)
}
