% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{porc_juveniles}
\alias{porc_juveniles}
\title{Porcentaje de juveniles}
\usage{
porc_juveniles(frecuencia, tallas, juvLim = 12)
}
\arguments{
\item{frecuencia}{Un vector numérico con la frecuencia de tallas muestreadas.}

\item{tallas}{Vector de tallas correspondiente a las frecuencias.}

\item{juvLim}{Talla límite (por defecto 12 cm) para clasificar como juvenil.}
}
\value{
Porcentaje de juveniles en la muestra.
}
\description{
Estima el porcentaje de individuos considerados juveniles en una muestra
según una talla límite establecida.
}
\examples{

frecuencia <- c(0, 1, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 10, 6, 3, 2)

tallas <- seq(5, 20, by = 1)

porc <- porc_juveniles(frecuencia, tallas, juvLim = 12)

print(porc)
}
