% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preparar_poligonos}
\alias{preparar_poligonos}
\title{Preparar polígonos a partir de datos de coordenadas}
\usage{
preparar_poligonos(datos, costa)
}
\arguments{
\item{datos}{Un data frame con coordenadas.}

\item{costa}{Un data frame con la línea de costa.}
}
\value{
Una lista de polígonos para visualización.
}
\description{
Función auxiliar para procesar datos y preparar polígonos.
}
\keyword{internal}
