% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helpers-announcement.R
\name{prepare_polygons}
\alias{prepare_polygons}
\title{Prepare polygons from coordinate data}
\usage{
prepare_polygons(
  data,
  coastline,
  coast_parallels = NULL,
  column_names = c("lat", "lon", "dc")
)
}
\arguments{
\item{data}{Data frame with coordinates.}

\item{coastline}{Data frame with coastline.}

\item{coast_parallels}{List of data frames with lines parallel to the coast.}

\item{column_names}{Vector with column names in coast_parallels data frames.}
}
\value{
List of polygons for visualization.
}
\description{
Helper function to process data and prepare polygons using preexisting parallel lines.
}
\keyword{internal}
