% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{process_json_response}
\alias{process_json_response}
\title{Procesar respuesta JSON}
\usage{
process_json_response(content_text, verbose = TRUE)
}
\arguments{
\item{content_text}{Texto de la respuesta}

\item{verbose}{Mostrar información detallada}
}
\value{
DataFrame con los anuncios o NULL si hay error
}
\description{
Procesa la respuesta JSON del API y extrae los anuncios
}
\keyword{internal}
