% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-refactor-data.R
\name{process_length}
\alias{process_length}
\title{Process length data from hauls}
\usage{
process_length(data_length, verbose = FALSE)
}
\arguments{
\item{data_length}{Data frame with raw length data.}

\item{verbose}{Logical. Print column mapping information.}
}
\value{
Data frame with length by haul in wide format with individual columns
for each length class.
}
\description{
Processes length data from PRODUCE logbooks. Automatically detects required
columns and transforms from long to wide format.
}
\examples{
length_data <- process_length(data_length = tallas_bitacora)
length_data <- process_length(data_length = tallas_bitacora, verbose = TRUE)

}
