% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{puntos_tierra}
\alias{puntos_tierra}
\title{Puntos en tierra}
\usage{
puntos_tierra(x_punto, y_punto, linea_costa, paralelo = FALSE, nucleos = 4)
}
\arguments{
\item{x_punto}{Un vector con las longitudes}

\item{y_punto}{Un vector con las latitudes}

\item{linea_costa}{Un data frame con las coordenadas de la costa, el orden debe ser: lon, lat. Default \code{Tivy::linea_costa_peru}.}

\item{paralelo}{Lógico. Si \code{TRUE}, realiza el cálculo en paralelo utilizando múltiples núcleos. Default \code{FALSE}.}

\item{nucleos}{Número de núcleos a usar para procesamiento paralelo. Default \code{4}.}
}
\value{
Un vector con clasificación "tierra" o "mar" para cada punto.
}
\description{
Función que permite estimar los puntos que están a la derecha o izquierda de la línea de costa
}
\examples{
data_calas = processing_calas(data_calas = calas)
puntos_tierra(x_punto = data_calas$lon_final, y_punto = data_calas$lat_final, linea_costa = Tivy::linea_costa_peru)
}
