% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tallas_bitacora}
\alias{tallas_bitacora}
\title{Length data sampled from hauls made by vessels in the Peruvian sea.}
\format{
A data.frame with sample data of lengths sampled from hauls made by vessels in the Peruvian sea.
}
\source{
Indicate the data source
}
\usage{
tallas_bitacora
}
\description{
Dataset containing information about lengths of marine species
captured along the Peruvian littoral.
}
\examples{
data(tallas_bitacora)
}
\keyword{datasets}
