% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tallas_bitacora}
\alias{tallas_bitacora}
\title{Datos de tallas de especies marinas}
\format{
Un data.frame con los siguientes atributos:
\describe{
\item{id}{Identificador único del registro}
\item{especie}{Nombre de la especie}
\item{talla}{Longitud o talla del individuo en cm o mm}
\item{fecha}{Fecha de captura o medición}
\item{ubicacion}{Ubicación donde se realizó la medición}
\item{...}{Otros atributos relevantes}
}
}
\source{
Indicar la fuente de los datos
}
\usage{
tallas_bitacora
}
\description{
Conjunto de datos que contiene información sobre tallas de especies marinas
capturadas en el litoral peruano.
}
\examples{
data(tallas_bitacora)
}
\keyword{datasets}
