% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fishery-statistics.R
\name{weight_by_catch}
\alias{weight_by_catch}
\title{Weight length frequencies by total catch}
\usage{
weight_by_catch(frequency, catch, length, a, b, silence_warnings = FALSE)
}
\arguments{
\item{frequency}{Numeric vector of observed length frequencies.}

\item{catch}{Total catch amount (in kg or tons).}

\item{length}{Numeric vector of lengths corresponding to frequencies.}

\item{a}{Coefficient of the length-weight relationship.}

\item{b}{Exponent of the length-weight relationship.}

\item{silence_warnings}{Logical. Suppress warning messages.}
}
\value{
Numeric vector of weighted frequencies.
}
\description{
Scales observed length frequencies based on total recorded catch using length-weight relationship.
}
\examples{
freq <- c(0, 4, 8, 16, 12, 23, 34, 55, 35, 24, 15, 6, 0)
lengths <- seq(8, 20, by = 1)
weighted_freq <- weight_by_catch(freq, catch = 1000, lengths, a = 0.0001, b = 2.984)

}
